Roll up v1.4 branch commit for bug email bugs 838843, 882917, 796474
Bug 838843 - [email] email search does not update as headers are modified or deleted, allows user to still see and fail to display deleted messages. r=mcav
land https://github.com/mozilla-b2g/gaia-email-libs-and-more/pull/296

Bug 882917 - [email] "Body" search of HTML messages regressed by worker thread migration. r=mcav
land https://github.com/mozilla-b2g/gaia-email-libs-and-more/pull/297

Bug 796474 - [email] quasi-infinite scroll may hitch under OMTC; consider using absolute positioning with virtual coordinate space r=asuth
