+++ apps/email/js/ext/mailapi/worker-bootstrap.js
@@ -5051,6 +5051,7 @@ exports.TEST_adjustSyncValues = function TEST_adjustSyncValues(syncValues) {
     fillSize: 'INITIAL_FILL_SIZE',
     days: 'INITIAL_SYNC_DAYS',
     growDays: 'INITIAL_SYNC_GROWTH_DAYS',
+    wholeFolderSync: 'SYNC_WHOLE_FOLDER_AT_N_MESSAGES',
     bisectThresh: 'BISECT_DATE_AT_N_MESSAGES',
     tooMany: 'TOO_MANY_MESSAGES',
     scaleFactor: 'TIME_SCALE_FACTOR_ON_NO_MESSAGES',
@@ -5206,8 +5207,8 @@ var SYNC_START_MINIMUM_PROGRESS = 0.02;
  * Headers are removed, added, or modified using the onHeader* methods.
  * The updates are sent to 'SliceBridgeProxy' which batches updates and
  * puts them on the event loop. We batch so that we can minimize the number of
- * reflows and painting on the DOM side. This also enables us to batch data 
- * received in network packets around the smae time without having to handle it in 
+ * reflows and painting on the DOM side. This also enables us to batch data
+ * received in network packets around the smae time without having to handle it in
  * each protocol's logic.
  *
  * Currently, we only batch updates that are done between 'now' and the next time
@@ -5249,9 +5250,16 @@ function MailSlice(bridgeHandle, storage, _parentLog) {
    */
   this.headers = [];
   this.desiredHeaders = $sync.INITIAL_FILL_SIZE;
+
+  this.headerCount = storage.headerCount;
 }
 exports.MailSlice = MailSlice;
 MailSlice.prototype = {
+  /**
+   * We are a folder-backed view-slice.
+   */
+  type: 'folder',
+
   set atTop(val) {
     if (this._bridgeHandle)
       this._bridgeHandle.atTop = val;
@@ -5270,6 +5278,11 @@ MailSlice.prototype = {
   set userCanGrowDownwards(val) {
     if (this._bridgeHandle)
       this._bridgeHandle.userCanGrowDownwards = val;
+    return val;
+  },
+  set headerCount(val) {
+    if (this._bridgeHandle)
+      this._bridgeHandle.headerCount = val;
     return val;
   },
 
@@ -5465,18 +5478,17 @@ MailSlice.prototype = {
    * called when the header is in the time-range of interest and a refresh,
    * cron-triggered sync, or IDLE/push tells us to do so.
    */
-  onHeaderAdded: function(header, syncDriven, messageIsNew) {
+  onHeaderAdded: function(header, body, syncDriven, messageIsNew) {
     if (!this._bridgeHandle)
       return;
 
     var idx = bsearchForInsert(this.headers, header, cmpHeaderYoungToOld);
     var hlen = this.headers.length;
-    // Don't append the header if it would expand us beyond our requested amount
-    // and there is no subsequent step, like accumulate flushing, that would get
-    // rid of the excess.  Note that this does not guarantee that we won't
-    // end up with more headers than originally planned; if we get told about
-    // headers earlier than the last slot, we will insert them and grow without
-    // forcing a removal of something else to offset.
+    // Don't append the header if it would expand us beyond our requested
+    // amount.  Note that this does not guarantee that we won't end up with more
+    // headers than originally planned; if we get told about headers earlier
+    // than the last slot, we will insert them and grow without forcing a
+    // removal of something else to offset.
     if (hlen >= this.desiredHeaders && idx === hlen)
       return;
     // If we are inserting (not at the end) then be sure to grow
@@ -5515,7 +5527,7 @@ MailSlice.prototype = {
    * Tells the slice that a header it should know about has changed.  (If
    * this is a search, it's okay for it not to know...)
    */
-  onHeaderModified: function(header) {
+  onHeaderModified: function(header, body) {
     if (!this._bridgeHandle)
       return;
 
@@ -5932,6 +5944,15 @@ function FolderStorage(account, folderId, persistedFolderInfo, dbConn,
    * }
    */
   this._headerBlockInfos = persistedFolderInfo.headerBlocks;
+
+  // Calculate total number of messages
+  this.headerCount = 0;
+  if (this._headerBlockInfos) {
+    this._headerBlockInfos.forEach(function(headerBlockInfo) {
+      this.headerCount += headerBlockInfo.count;
+    }.bind(this));
+  }
+
   /**
    * @listof[FolderBlockInfo]{
    *   Newest-to-oldest (numerically decreasing time and ID) sorted list of
@@ -5985,6 +6006,9 @@ function FolderStorage(account, folderId, persistedFolderInfo, dbConn,
    */
   this._loadedBodyBlockInfos = [];
 
+  this._flushExcessTimeoutId = 0;
+
+  this._bound_flushExcessOnTimeout = this._flushExcessOnTimeout.bind(this);
   this._bound_makeHeaderBlock = this._makeHeaderBlock.bind(this);
   this._bound_insertHeaderInBlock = this._insertHeaderInBlock.bind(this);
   this._bound_splitHeaderBlock = this._splitHeaderBlock.bind(this);
@@ -5994,6 +6018,7 @@ function FolderStorage(account, folderId, persistedFolderInfo, dbConn,
   this._bound_insertBodyInBlock = this._insertBodyInBlock.bind(this);
   this._bound_splitBodyBlock = this._splitBodyBlock.bind(this);
   this._bound_deleteBodyFromBlock = this._deleteBodyFromBlock.bind(this);
+
 
   /**
    * Has our internal state altered at all and will need to be persisted?
@@ -6041,9 +6066,10 @@ function FolderStorage(account, folderId, persistedFolderInfo, dbConn,
   this._mutexQueue = [];
 
   /**
-   * Active view slices on this folder.
+   * Active view / search slices on this folder.
    */
   this._slices = [];
+
   /**
    * The slice that is driving our current synchronization and wants to hear
    * about all header modifications/notes as they occur.  This will be null
@@ -6072,9 +6098,11 @@ FolderStorage.prototype = {
     // will occur, but we don't care.)
     for (var i = this._slices.length - 1; i >= 0; i--) {
       var slice = this._slices[i];
-      slice.reset();
       slice.desiredHeaders = $sync.INITIAL_FILL_SIZE;
-      this._resetAndResyncSlice(slice, true, null);
+      slice.reset();
+      if (slice.type === 'folder') {
+        this._resetAndResyncSlice(slice, true, null);
+      }
     }
   },
 
@@ -6437,16 +6465,26 @@ FolderStorage.prototype = {
    * Flush cached blocks that are unlikely to be used again soon.  Our
    * heuristics for deciding what to keep is simple:
    * - Dirty blocks are always kept; this is required for correctness.
-   * - Blocks that overlap with live `MailSlice` instances are kept.
+   * - Header blocks that overlap with live `MailSlice` instances are kept.
    *
    * It could also make sense to support some type of MRU tracking, but the
    * complexity is not currently justified since the live `MailSlice` should
    * lead to a near-perfect hit rate on immediate actions and the UI's
    * pre-emptive slice growing should insulate it from any foolish discards
    * we might make.
+   *
+   * For bodies, since they are larger, and the UI may not always shrink a
+   * slice, only keep around one blockInfo of them, which contain the most
+   * likely immediately needed blockInfos, for instance a direction reversal
+   * in a next/previous navigation.
    */
   flushExcessCachedBlocks: function(debugLabel) {
-    var slices = this._slices;
+    // We only care about explicitly folder-backed slices for cache eviction
+    // purposes.  Search filters are sparse and would keep way too much in
+    // memory.
+    var slices = this._slices.filter(function (slice) {
+                   return slice.type === 'folder';
+                 });
     function blockIntersectsAnySlice(blockInfo) {
       for (var i = 0; i < slices.length; i++) {
         var slice = slices[i];
@@ -6465,30 +6503,72 @@ FolderStorage.prototype = {
       return false;
     }
     function maybeDiscard(blockType, blockInfoList, loadedBlockInfos,
-                          blockMap, dirtyMap) {
+                          blockMap, dirtyMap, shouldDiscardFunc) {
       // console.warn('!! flushing', blockType, 'blocks because:', debugLabel);
-      for (var i = 0; i < loadedBlockInfos.length; i++) {
+
+      // Go backwards in array, to allow code to keep a count of
+      // blockInfos to keep that favor the most current ones.
+      for (var i = loadedBlockInfos.length - 1; i > -1; i--) {
         var blockInfo = loadedBlockInfos[i];
         // do not discard dirty blocks
         if (dirtyMap.hasOwnProperty(blockInfo.blockId)) {
           // console.log('  dirty block:', blockInfo.blockId);
           continue;
         }
-        // do not discard blocks that overlap mail slices
-        if (blockIntersectsAnySlice(blockInfo))
-          continue;
-        // console.log('discarding', blockType, 'block', blockInfo.blockId);
-        delete blockMap[blockInfo.blockId];
-        loadedBlockInfos.splice(i--, 1);
+
+        if (shouldDiscardFunc(blockInfo)) {
+          // console.log('discarding', blockType, 'block', blockInfo.blockId);
+          delete blockMap[blockInfo.blockId];
+          loadedBlockInfos.splice(i, 1);
+        }
       }
     }
 
     maybeDiscard(
-      'header', this._headerBlockInfos, this._loadedHeaderBlockInfos,
-      this._headerBlocks, this._dirtyHeaderBlocks);
+      'header',
+      this._headerBlockInfos,
+      this._loadedHeaderBlockInfos,
+      this._headerBlocks,
+      this._dirtyHeaderBlocks,
+      function (blockInfo) {
+        // Do not discard blocks that overlap mail slices.
+        return !blockIntersectsAnySlice(blockInfo);
+      }
+    );
+
+    // Keep one body block around if there are open folder slices.  If there are
+    // no open slices, discard everything.  (If there are no headers then there
+    // isn't really a way to access the bodies.)
+    var keepCount = slices.length ? 1 : 0,
+        foundCount = 0;
+
     maybeDiscard(
-      'body', this._bodyBlockInfos, this._loadedBodyBlockInfos,
-      this._bodyBlocks, this._dirtyBodyBlocks);
+      'body',
+      this._bodyBlockInfos,
+      this._loadedBodyBlockInfos,
+      this._bodyBlocks,
+      this._dirtyBodyBlocks,
+      function(blockInfo) {
+        // For bodies, want to always purge as front end may decide to
+        // never shrink a messages slice, but keep one block around to
+        // avoid wasteful DB IO for commonly grouped operations, for
+        // example, a next/previous message navigation direction change.
+        foundCount += 1;
+        return foundCount > keepCount;
+      }
+    );
+  },
+
+  /**
+   * Called after a timeout to do cleanup of cached blocks to keep memory
+   * low. However, only do the cleanup if there is no more mutex-controlled
+   * work so as to keep likely useful cache items still in memory.
+   */
+  _flushExcessOnTimeout: function() {
+    this._flushExcessTimeoutId = 0;
+    if (!this.isDead && this._mutexQueue.length === 0) {
+      this.flushExcessCachedBlocks('flushExcessOnTimeout');
+    }
   },
 
   /**
@@ -7122,11 +7202,24 @@ FolderStorage.prototype = {
       this._loadBlock(type, info, processBlock.bind(this));
   },
 
-  runAfterDeferredCalls: function(callback) {
-    if (this._deferredCalls.length)
+  /**
+   * Run the given callback after all pending deferred calls have run.
+   *
+   * @param {Function} callback
+   * @param {Boolean} [alwaysDefer=false]
+   *   Should we defer the callback to the next turn of the event loop even
+   *   if there's no reason to wait?  Arguably this is what we should always
+   *   do (at least by default) for human sanity purposes, but existing code
+   *   would need to be audited.
+   */
+  runAfterDeferredCalls: function(callback, alwaysDefer) {
+    if (this._deferredCalls.length) {
       this._deferredCalls.push(callback);
-    else
+    } else if (alwaysDefer) {
+      window.setZeroTimeout(callback);
+    } else {
       callback();
+    }
   },
 
   /**
@@ -7203,6 +7296,19 @@ FolderStorage.prototype = {
 
       if (self._pendingLoads.length === 0)
         self._runDeferredCalls();
+
+      // Ask for cleanup of old blocks in case the UI is not shrinking
+      // any slices.
+      if (self._mutexQueue.length === 0 && !self._flushExcessTimeoutId) {
+        self._flushExcessTimeoutId = setTimeout(
+          self._bound_flushExcessOnTimeout,
+          // Choose 5 seconds, since it is a human-scale value around
+          // the order of how long we expect it would take the user
+          // to realize they hit the opposite arrow navigation button
+          // from what they meant.
+          5000
+        );
+      }
     }
 
     this._LOG.loadBlock_begin(type, blockId);
@@ -7262,6 +7368,10 @@ FolderStorage.prototype = {
       processBlock.call(this, blockMap[info.blockId]);
     else
       this._loadBlock(type, info, processBlock.bind(this));
+  },
+
+  sliceOpenSearch: function fs_sliceOpenSearch(slice) {
+    this._slices.push(slice);
   },
 
   /**
@@ -7663,6 +7773,15 @@ FolderStorage.prototype = {
         doneCallback, progressCallback);
     }.bind(this);
 
+    // The front end may not be calling shrink any more, to reduce
+    // complexity for virtual scrolling. So be sure to clear caches
+    // that are not needed, to avoid a large memory growth from
+    // keeping the header bodies as the user does next/previous
+    // navigation.
+    if (this._mutexQueue.length === 0) {
+      this.flushExcessCachedBlocks('grow');
+    }
+
     // --- request messages
     if (dirMagnitude < 0) {
       dir = FUTUREWARDS;
@@ -7854,8 +7973,15 @@ FolderStorage.prototype = {
     var idx = this._slices.indexOf(slice);
     this._slices.splice(idx, 1);
 
-    if (this._slices.length === 0 && this._mutexQueue.length === 0)
+    // If this was a folder-backed slice, we potentially can now free up a lot
+    // of cached memory, so do that.
+    if (slice.type === 'folder') {
+      this.flushExcessCachedBlocks('deadslice');
+    }
+
+    if (this._slices.length === 0 && this._mutexQueue.length === 0) {
       this.folderSyncer.allConsumersDead();
+    }
   },
 
   /**
@@ -8703,21 +8829,36 @@ FolderStorage.prototype = {
 
   /**
    * Add a new message to the database, generating slice notifications.
+   *
+   * @param header
+   * @param [body]
+   *   Optional body, exists to hint to slices so that SearchFilter can peek
+   *   directly at the body without needing to make an additional request to
+   *   look at the body.
    */
-  addMessageHeader: function ifs_addMessageHeader(header, callback) {
+  addMessageHeader: function ifs_addMessageHeader(header, body, callback) {
     if (header.id == null || header.suid == null) {
       throw new Error('No valid id: ' + header.id + ' or suid: ' + header.suid);
     }
 
     if (this._pendingLoads.length) {
       this._deferredCalls.push(this.addMessageHeader.bind(
-                                 this, header, callback));
+                                 this, header, body, callback));
       return;
     }
     this._LOG.addMessageHeader(header.date, header.id, header.srvid);
 
-    if (this._curSyncSlice && !this._curSyncSlice.ignoreHeaders)
-      this._curSyncSlice.onHeaderAdded(header, true, true);
+    this.headerCount += 1;
+
+    if (this._curSyncSlice && !this._curSyncSlice.ignoreHeaders) {
+      // TODO: make sure the slice knows the true offset of its
+      // first header in the folder. Currently the UI never
+      // shrinks its slice so this number is always 0 and we can
+      // get away without providing that offset for now.
+      this._curSyncSlice.headerCount = this.headerCount;
+      this._curSyncSlice.onHeaderAdded(header, body, true, true);
+    }
+
     // - Generate notifications for (other) interested slices
     if (this._slices.length > (this._curSyncSlice ? 1 : 0)) {
       var date = header.date, uid = header.id;
@@ -8761,10 +8902,29 @@ FolderStorage.prototype = {
           slice.desiredHeaders++;
         }
 
-        if (slice._onAddingHeader)
-          slice._onAddingHeader(header);
+        if (slice.type === 'folder') {
+          // TODO: make sure the slice knows the true offset of its
+          // first header in the folder. Currently the UI never
+          // shrinks its slice so this number is always 0 and we can
+          // get away without providing that offset for now.
+          slice.headerCount = this.headerCount;
+        }
 
-        slice.onHeaderAdded(header, false, true);
+        if (slice._onAddingHeader) {
+          try {
+            slice._onAddingHeader(header);
+          }
+          catch (ex) {
+            this._LOG.callbackErr(ex);
+          }
+        }
+
+        try {
+          slice.onHeaderAdded(header, body, false, true);
+        }
+        catch (ex) {
+          this._LOG.callbackErr(ex);
+        }
       }
     }
 
@@ -8788,13 +8948,14 @@ FolderStorage.prototype = {
    */
   updateMessageHeader: function ifs_updateMessageHeader(date, id, partOfSync,
                                                         headerOrMutationFunc,
+                                                        body,
                                                         callback) {
     // (While this method can complete synchronously, we want to maintain its
     // perceived ordering relative to those that cannot be.)
     if (this._pendingLoads.length) {
       this._deferredCalls.push(this.updateMessageHeader.bind(
                                  this, date, id, partOfSync,
-                                 headerOrMutationFunc, callback));
+                                 headerOrMutationFunc, body, callback));
       return;
     }
 
@@ -8844,7 +9005,12 @@ FolderStorage.prototype = {
                 (date === slice.endTS &&
                  id > slice.endUID))
               continue;
-            slice.onHeaderModified(header);
+            try {
+              slice.onHeaderModified(header, body);
+            }
+            catch (ex) {
+              this._LOG.callbackErr(ex);
+            }
           }
         }
       }
@@ -8868,7 +9034,7 @@ FolderStorage.prototype = {
    * Retrieve and update a header by locating it
    */
   updateMessageHeaderByServerId: function(srvid, partOfSync,
-                                          headerOrMutationFunc) {
+                                          headerOrMutationFunc, body) {
     if (this._pendingLoads.length) {
       this._deferredCalls.push(this.updateMessageHeaderByServerId.bind(
         this, srvid, partOfSync, headerOrMutationFunc));
@@ -8889,7 +9055,7 @@ FolderStorage.prototype = {
           // future work: this method will duplicate some work to re-locate
           // the header; we could try and avoid doing that.
           this.updateMessageHeader(
-            header.date, header.id, partOfSync, headerOrMutationFunc);
+            header.date, header.id, partOfSync, headerOrMutationFunc, body);
           return;
         }
       }
@@ -8946,11 +9112,28 @@ FolderStorage.prototype = {
       return;
     }
 
-    if (this._curSyncSlice && !this._curSyncSlice.ignoreHeaders)
+    this.headerCount -= 1;
+
+    if (this._curSyncSlice && !this._curSyncSlice.ignoreHeaders) {
+      // TODO: make sure the slice knows the true offset of its
+      // first header in the folder. Currently the UI never
+      // shrinks its slice so this number is always 0 and we can
+      // get away without providing that offset for now.
+      this._curSyncSlice.headerCount = this.headerCount;
       this._curSyncSlice.onHeaderRemoved(header);
+    }
     if (this._slices.length > (this._curSyncSlice ? 1 : 0)) {
       for (var iSlice = 0; iSlice < this._slices.length; iSlice++) {
         var slice = this._slices[iSlice];
+
+        if (slice.type === 'folder') {
+          // TODO: make sure the slice knows the true offset of its
+          // first header in the folder. Currently the UI never
+          // shrinks its slice so this number is always 0 and we can
+          // get away without providing that offset for now.
+          slice.headerCount = this.headerCount;
+        }
+
         if (slice === this._curSyncSlice)
           continue;
         if (BEFORE(header.date, slice.startTS) ||
@@ -8961,6 +9144,7 @@ FolderStorage.prototype = {
             (header.date === slice.endTS &&
              header.id > slice.endUID))
           continue;
+
         slice.onHeaderRemoved(header);
       }
     }
@@ -9033,7 +9217,7 @@ FolderStorage.prototype = {
                                  this, header, bodyInfo, callback));
       return;
     }
-    this._LOG.addMessageBody(header.date, header.id, header.srvid);
+    this._LOG.addMessageBody(header.date, header.id, header.srvid, bodyInfo);
 
     // crappy size estimates where we assume the world is ASCII and so a UTF-8
     // encoding will take exactly 1 byte per character.
@@ -9160,7 +9344,22 @@ FolderStorage.prototype = {
     });
   },
 
+  /**
+   * Load the given message body while obeying call ordering consistency rules.
+   * If any other calls have gone asynchronous because block loads are required,
+   * then this call will wait for those calls to complete first even if we
+   * already have the requested body block loaded.  If we haven't gone async and
+   * the body is already available, the callback will be invoked synchronously
+   * while this function is still on the stack.  So, uh, don't be surprised by
+   * that.
+   */
   getMessageBody: function ifs_getMessageBody(suid, date, callback) {
+    if (this._pendingLoads.length) {
+      this._deferredCalls.push(
+        this.getMessageBody.bind(this, suid, date, callback));
+      return;
+    }
+
     var id = parseInt(suid.substring(suid.lastIndexOf('/') + 1)),
         posInfo = this._findRangeObjIndexForDateAndID(this._bodyBlockInfos,
                                                       date, id);
@@ -9410,6 +9609,7 @@ var LOGFAB = exports.LOGFAB = $log.register($module, {
       syncedToDawnOfTime: {},
     },
     TEST_ONLY_events: {
+      addMessageBody: { body: false }
     },
     asyncJobs: {
       loadBlock: { type: false, blockId: false },
@@ -9731,6 +9931,2770 @@ return {
 
 }); // end define
 ;
+(function (root, factory) {
+    // Universal Module Definition (UMD) to support AMD, CommonJS/Node.js,
+    // Rhino, and plain browser loading.
+    if (typeof define === 'function' && define.amd) {
+        define('bleach/css-parser/tokenizer',['exports'], factory);
+    } else if (typeof exports !== 'undefined') {
+        factory(exports);
+    } else {
+        factory(root);
+    }
+}(this, function (exports) {
+
+var between = function (num, first, last) { return num >= first && num <= last; }
+function digit(code) { return between(code, 0x30,0x39); }
+function hexdigit(code) { return digit(code) || between(code, 0x41,0x46) || between(code, 0x61,0x66); }
+function uppercaseletter(code) { return between(code, 0x41,0x5a); }
+function lowercaseletter(code) { return between(code, 0x61,0x7a); }
+function letter(code) { return uppercaseletter(code) || lowercaseletter(code); }
+function nonascii(code) { return code >= 0xa0; }
+function namestartchar(code) { return letter(code) || nonascii(code) || code == 0x5f; }
+function namechar(code) { return namestartchar(code) || digit(code) || code == 0x2d; }
+function nonprintable(code) { return between(code, 0,8) || between(code, 0xe,0x1f) || between(code, 0x7f,0x9f); }
+function newline(code) { return code == 0xa || code == 0xc; }
+function whitespace(code) { return newline(code) || code == 9 || code == 0x20; }
+function badescape(code) { return newline(code) || isNaN(code); }
+
+// Note: I'm not yet acting smart enough to actually handle astral characters.
+var maximumallowedcodepoint = 0x10ffff;
+
+function tokenize(str, options) {
+	if(options == undefined) options = {transformFunctionWhitespace:false, scientificNotation:false};
+	var i = -1;
+	var tokens = [];
+	var state = "data";
+	var code;
+	var currtoken;
+
+	// Line number information.
+	var line = 0;
+	var column = 0;
+	// The only use of lastLineLength is in reconsume().
+	var lastLineLength = 0;
+	var incrLineno = function() {
+		line += 1;
+		lastLineLength = column;
+		column = 0;
+	};
+	var locStart = {line:line, column:column};
+
+	var next = function(num) { if(num === undefined) num = 1; return str.charCodeAt(i+num); };
+	var consume = function(num) {
+		if(num === undefined)
+			num = 1;
+		i += num;
+		code = str.charCodeAt(i);
+		if (newline(code)) incrLineno();
+		else column += num;
+		//console.log('Consume '+i+' '+String.fromCharCode(code) + ' 0x' + code.toString(16));
+		return true;
+	};
+	var reconsume = function() {
+		i -= 1;
+		if (newline(code)) {
+			line -= 1;
+			column = lastLineLength;
+		} else {
+			column -= 1;
+		}
+		locStart.line = line;
+		locStart.column = column;
+		return true;
+	};
+	var eof = function() { return i >= str.length; };
+	var donothing = function() {};
+	var emit = function(token) {
+		if(token) {
+			token.finish();
+		} else {
+			token = currtoken.finish();
+		}
+		if (options.loc === true) {
+			token.loc = {};
+			token.loc.start = {line:locStart.line, column:locStart.column, idx: locStart.idx};
+			locStart = {line: line, column: column, idx: i};
+			token.loc.end = locStart;
+		}
+		tokens.push(token);
+		//console.log('Emitting ' + token);
+		currtoken = undefined;
+		return true;
+	};
+	var create = function(token) { currtoken = token; return true; };
+	var parseerror = function() { console.log("Parse error at index " + i + ", processing codepoint 0x" + code.toString(16) + " in state " + state + ".");return true; };
+	var catchfire = function(msg) { console.log("MAJOR SPEC ERROR: " + msg); return true;}
+	var switchto = function(newstate) {
+		state = newstate;
+		//console.log('Switching to ' + state);
+		return true;
+	};
+	var consumeEscape = function() {
+		// Assume the the current character is the \
+		consume();
+		if(hexdigit(code)) {
+			// Consume 1-6 hex digits
+			var digits = [];
+			for(var total = 0; total < 6; total++) {
+				if(hexdigit(code)) {
+					digits.push(code);
+					consume();
+				} else { break; }
+			}
+			var value = parseInt(digits.map(String.fromCharCode).join(''), 16);
+			if( value > maximumallowedcodepoint ) value = 0xfffd;
+			// If the current char is whitespace, cool, we'll just eat it.
+			// Otherwise, put it back.
+			if(!whitespace(code)) reconsume();
+			return value;
+		} else {
+			return code;
+		}
+	};
+
+	for(;;) {
+		if(i > str.length*2) return "I'm infinite-looping!";
+		consume();
+		switch(state) {
+		case "data":
+			if(whitespace(code)) {
+				emit(new WhitespaceToken);
+				while(whitespace(next())) consume();
+			}
+			else if(code == 0x22) switchto("double-quote-string");
+			else if(code == 0x23) switchto("hash");
+			else if(code == 0x27) switchto("single-quote-string");
+			else if(code == 0x28) emit(new OpenParenToken);
+			else if(code == 0x29) emit(new CloseParenToken);
+			else if(code == 0x2b) {
+				if(digit(next()) || (next() == 0x2e && digit(next(2)))) switchto("number") && reconsume();
+				else emit(new DelimToken(code));
+			}
+			else if(code == 0x2d) {
+				if(next(1) == 0x2d && next(2) == 0x3e) consume(2) && emit(new CDCToken);
+				else if(digit(next()) || (next(1) == 0x2e && digit(next(2)))) switchto("number") && reconsume();
+				else switchto('ident') && reconsume();
+			}
+			else if(code == 0x2e) {
+				if(digit(next())) switchto("number") && reconsume();
+				else emit(new DelimToken(code));
+			}
+			else if(code == 0x2f) {
+				if(next() == 0x2a) consume() && switchto("comment");
+				else emit(new DelimToken(code));
+			}
+			else if(code == 0x3a) emit(new ColonToken);
+			else if(code == 0x3b) emit(new SemicolonToken);
+			else if(code == 0x3c) {
+				if(next(1) == 0x21 && next(2) == 0x2d && next(3) == 0x2d) consume(3) && emit(new CDOToken);
+				else emit(new DelimToken(code));
+			}
+			else if(code == 0x40) switchto("at-keyword");
+			else if(code == 0x5b) emit(new OpenSquareToken);
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit(new DelimToken(code));
+				else switchto('ident') && reconsume();
+			}
+			else if(code == 0x5d) emit(new CloseSquareToken);
+			else if(code == 0x7b) emit(new OpenCurlyToken);
+			else if(code == 0x7d) emit(new CloseCurlyToken);
+			else if(digit(code)) switchto("number") && reconsume();
+			else if(code == 0x55 || code == 0x75) {
+				if(next(1) == 0x2b && hexdigit(next(2))) consume() && switchto("unicode-range");
+				else switchto('ident') && reconsume();
+			}
+			else if(namestartchar(code)) switchto('ident') && reconsume();
+			else if(eof()) { emit(new EOFToken); return tokens; }
+			else emit(new DelimToken(code));
+			break;
+
+		case "double-quote-string":
+			if(currtoken == undefined) create(new StringToken);
+
+			if(code == 0x22) emit() && switchto("data");
+			else if(eof()) parseerror() && emit() && switchto("data") && reconsume();
+			else if(newline(code)) parseerror() && emit(new BadStringToken) && switchto("data") && reconsume();
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit(new BadStringToken) && switchto("data");
+				else if(newline(next())) consume();
+				else currtoken.append(consumeEscape());
+			}
+			else currtoken.append(code);
+			break;
+
+		case "single-quote-string":
+			if(currtoken == undefined) create(new StringToken);
+
+			if(code == 0x27) emit() && switchto("data");
+			else if(eof()) parseerror() && emit() && switchto("data");
+			else if(newline(code)) parseerror() && emit(new BadStringToken) && switchto("data") && reconsume();
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit(new BadStringToken) && switchto("data");
+				else if(newline(next())) consume();
+				else currtoken.append(consumeEscape());
+			}
+			else currtoken.append(code);
+			break;
+
+		case "hash":
+			if(namechar(code)) create(new HashToken(code)) && switchto("hash-rest");
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit(new DelimToken(0x23)) && switchto("data") && reconsume();
+				else create(new HashToken(consumeEscape())) && switchto('hash-rest');
+			}
+			else emit(new DelimToken(0x23)) && switchto('data') && reconsume();
+			break;
+
+		case "hash-rest":
+			if(namechar(code)) currtoken.append(code);
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit() && switchto("data") && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "comment":
+			if(code == 0x2a) {
+				if(next() == 0x2f) consume() && switchto('data');
+				else donothing();
+			}
+			else if(eof()) parseerror() && switchto('data') && reconsume();
+			else donothing();
+			break;
+
+		case "at-keyword":
+			if(code == 0x2d) {
+				if(namestartchar(next())) create(new AtKeywordToken(0x2d)) && switchto('at-keyword-rest');
+				else if(next(1) == 0x5c && !badescape(next(2))) create(new AtKeywordtoken(0x2d)) && switchto('at-keyword-rest');
+				else parseerror() && emit(new DelimToken(0x40)) && switchto('data') && reconsume();
+			}
+			else if(namestartchar(code)) create(new AtKeywordToken(code)) && switchto('at-keyword-rest');
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit(new DelimToken(0x23)) && switchto("data") && reconsume();
+				else create(new AtKeywordToken(consumeEscape())) && switchto('at-keyword-rest');
+			}
+			else emit(new DelimToken(0x40)) && switchto('data') && reconsume();
+			break;
+
+		case "at-keyword-rest":
+			if(namechar(code)) currtoken.append(code);
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit() && switchto("data") && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "ident":
+			if(code == 0x2d) {
+				if(namestartchar(next())) create(new IdentifierToken(code)) && switchto('ident-rest');
+				else if(next(1) == 0x5c && !badescape(next(2))) create(new IdentifierToken(code)) && switchto('ident-rest');
+				else emit(new DelimToken(0x2d)) && switchto('data');
+			}
+			else if(namestartchar(code)) create(new IdentifierToken(code)) && switchto('ident-rest');
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && switchto("data") && reconsume();
+				else create(new IdentifierToken(consumeEscape())) && switchto('ident-rest');
+			}
+			else catchfire("Hit the generic 'else' clause in ident state.") && switchto('data') && reconsume();
+			break;
+
+		case "ident-rest":
+			if(namechar(code)) currtoken.append(code);
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit() && switchto("data") && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else if(code == 0x28) {
+				if(currtoken.ASCIImatch('url')) switchto('url');
+				else emit(new FunctionToken(currtoken)) && switchto('data');
+			}
+			else if(whitespace(code) && options.transformFunctionWhitespace) switchto('transform-function-whitespace') && reconsume();
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "transform-function-whitespace":
+			if(whitespace(next())) donothing();
+			else if(code == 0x28) emit(new FunctionToken(currtoken)) && switchto('data');
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "number":
+			create(new NumberToken());
+
+			if(code == 0x2d) {
+				if(digit(next())) consume() && currtoken.append([0x2d,code]) && switchto('number-rest');
+				else if(next(1) == 0x2e && digit(next(2))) consume(2) && currtoken.append([0x2d,0x2e,code]) && switchto('number-fraction');
+				else switchto('data') && reconsume();
+			}
+			else if(code == 0x2b) {
+				if(digit(next())) consume() && currtoken.append([0x2b,code]) && switchto('number-rest');
+				else if(next(1) == 0x2e && digit(next(2))) consume(2) && currtoken.append([0x2b,0x2e,code]) && switchto('number-fraction');
+				else switchto('data') && reconsume();
+			}
+			else if(digit(code)) currtoken.append(code) && switchto('number-rest');
+			else if(code == 0x2e) {
+				if(digit(next())) consume() && currtoken.append([0x2e,code]) && switchto('number-fraction');
+				else switchto('data') && reconsume();
+			}
+			else switchto('data') && reconsume();
+			break;
+
+		case "number-rest":
+			if(digit(code)) currtoken.append(code);
+			else if(code == 0x2e) {
+				if(digit(next())) consume() && currtoken.append([0x2e,code]) && switchto('number-fraction');
+				else emit() && switchto('data') && reconsume();
+			}
+			else if(code == 0x25) emit(new PercentageToken(currtoken)) && switchto('data');
+			else if(code == 0x45 || code == 0x65) {
+				if(digit(next())) consume() && currtoken.append([0x25,code]) && switchto('sci-notation');
+				else if((next(1) == 0x2b || next(1) == 0x2d) && digit(next(2))) currtoken.append([0x25,next(1),next(2)]) && consume(2) && switchto('sci-notation');
+				else create(new DimensionToken(currtoken,code)) && switchto('dimension');
+			}
+			else if(code == 0x2d) {
+				if(namestartchar(next())) consume() && create(new DimensionToken(currtoken,[0x2d,code])) && switchto('dimension');
+				else if(next(1) == 0x5c && badescape(next(2))) parseerror() && emit() && switchto('data') && reconsume();
+				else if(next(1) == 0x5c) consume() && create(new DimensionToken(currtoken, [0x2d,consumeEscape()])) && switchto('dimension');
+				else emit() && switchto('data') && reconsume();
+			}
+			else if(namestartchar(code)) create(new DimensionToken(currtoken, code)) && switchto('dimension');
+			else if(code == 0x5c) {
+				if(badescape(next)) parseerror() && emit() && switchto('data') && reconsume();
+				else create(new DimensionToken(currtoken,consumeEscape)) && switchto('dimension');
+			}
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "number-fraction":
+			currtoken.type = "number";
+
+			if(digit(code)) currtoken.append(code);
+			else if(code == 0x25) emit(new PercentageToken(currtoken)) && switchto('data');
+			else if(code == 0x45 || code == 0x65) {
+				if(digit(next())) consume() && currtoken.append([0x65,code]) && switchto('sci-notation');
+				else if((next(1) == 0x2b || next(1) == 0x2d) && digit(next(2))) currtoken.append([0x65,next(1),next(2)]) && consume(2) && switchto('sci-notation');
+				else create(new DimensionToken(currtoken,code)) && switchto('dimension');
+			}
+			else if(code == 0x2d) {
+				if(namestartchar(next())) consume() && create(new DimensionToken(currtoken,[0x2d,code])) && switchto('dimension');
+				else if(next(1) == 0x5c && badescape(next(2))) parseerror() && emit() && switchto('data') && reconsume();
+				else if(next(1) == 0x5c) consume() && create(new DimensionToken(currtoken, [0x2d,consumeEscape()])) && switchto('dimension');
+				else emit() && switchto('data') && reconsume();
+			}
+			else if(namestartchar(code)) create(new DimensionToken(currtoken, code)) && switchto('dimension');
+			else if(code == 0x5c) {
+				if(badescape(next)) parseerror() && emit() && switchto('data') && reconsume();
+				else create(new DimensionToken(currtoken,consumeEscape())) && switchto('dimension');
+			}
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "dimension":
+			if(namechar(code)) currtoken.append(code);
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && emit() && switchto('data') && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "sci-notation":
+			currtoken.type = "number";
+
+			if(digit(code)) currtoken.append(code);
+			else emit() && switchto('data') && reconsume();
+			break;
+
+		case "url":
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(code == 0x22) switchto('url-double-quote');
+			else if(code == 0x27) switchto('url-single-quote');
+			else if(code == 0x29) emit(new URLToken) && switchto('data');
+			else if(whitespace(code)) donothing();
+			else switchto('url-unquoted') && reconsume();
+			break;
+
+		case "url-double-quote":
+			if(! (currtoken instanceof URLToken)) create(new URLToken);
+
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(code == 0x22) switchto('url-end');
+			else if(newline(code)) parseerror() && switchto('bad-url');
+			else if(code == 0x5c) {
+				if(newline(next())) consume();
+				else if(badescape(next())) parseerror() && emit(new BadURLToken) && switchto('data') && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else currtoken.append(code);
+			break;
+
+		case "url-single-quote":
+			if(! (currtoken instanceof URLToken)) create(new URLToken);
+
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(code == 0x27) switchto('url-end');
+			else if(newline(code)) parseerror() && switchto('bad-url');
+			else if(code == 0x5c) {
+				if(newline(next())) consume();
+				else if(badescape(next())) parseerror() && emit(new BadURLToken) && switchto('data') && reconsume();
+				else currtoken.append(consumeEscape());
+			}
+			else currtoken.append(code);
+			break;
+
+		case "url-end":
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(whitespace(code)) donothing();
+			else if(code == 0x29) emit() && switchto('data');
+			else parseerror() && switchto('bad-url') && reconsume();
+			break;
+
+		case "url-unquoted":
+			if(! (currtoken instanceof URLToken)) create(new URLToken);
+
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(whitespace(code)) switchto('url-end');
+			else if(code == 0x29) emit() && switchto('data');
+			else if(code == 0x22 || code == 0x27 || code == 0x28 || nonprintable(code)) parseerror() && switchto('bad-url');
+			else if(code == 0x5c) {
+				if(badescape(next())) parseerror() && switchto('bad-url');
+				else currtoken.append(consumeEscape());
+			}
+			else currtoken.append(code);
+			break;
+
+		case "bad-url":
+			if(eof()) parseerror() && emit(new BadURLToken) && switchto('data');
+			else if(code == 0x29) emit(new BadURLToken) && switchto('data');
+			else if(code == 0x5c) {
+				if(badescape(next())) donothing();
+				else consumeEscape();
+			}
+			else donothing();
+			break;
+
+		case "unicode-range":
+			// We already know that the current code is a hexdigit.
+
+			var start = [code], end = [code];
+
+			for(var total = 1; total < 6; total++) {
+				if(hexdigit(next())) {
+					consume();
+					start.push(code);
+					end.push(code);
+				}
+				else break;
+			}
+
+			if(next() == 0x3f) {
+				for(;total < 6; total++) {
+					if(next() == 0x3f) {
+						consume();
+						start.push("0".charCodeAt(0));
+						end.push("f".charCodeAt(0));
+					}
+					else break;
+				}
+				emit(new UnicodeRangeToken(start,end)) && switchto('data');
+			}
+			else if(next(1) == 0x2d && hexdigit(next(2))) {
+				consume();
+				consume();
+				end = [code];
+				for(var total = 1; total < 6; total++) {
+					if(hexdigit(next())) {
+						consume();
+						end.push(code);
+					}
+					else break;
+				}
+				emit(new UnicodeRangeToken(start,end)) && switchto('data');
+			}
+			else emit(new UnicodeRangeToken(start)) && switchto('data');
+			break;
+
+		default:
+			catchfire("Unknown state '" + state + "'");
+		}
+	}
+}
+
+function stringFromCodeArray(arr) {
+	return String.fromCharCode.apply(null,arr.filter(function(e){return e;}));
+}
+
+function CSSParserToken(options) { return this; }
+CSSParserToken.prototype.finish = function() { return this; }
+CSSParserToken.prototype.toString = function() { return this.tokenType; }
+CSSParserToken.prototype.toJSON = function() { return this.toString(); }
+
+function BadStringToken() { return this; }
+BadStringToken.prototype = new CSSParserToken;
+BadStringToken.prototype.tokenType = "BADSTRING";
+
+function BadURLToken() { return this; }
+BadURLToken.prototype = new CSSParserToken;
+BadURLToken.prototype.tokenType = "BADURL";
+
+function WhitespaceToken() { return this; }
+WhitespaceToken.prototype = new CSSParserToken;
+WhitespaceToken.prototype.tokenType = "WHITESPACE";
+WhitespaceToken.prototype.toString = function() { return "WS"; }
+
+function CDOToken() { return this; }
+CDOToken.prototype = new CSSParserToken;
+CDOToken.prototype.tokenType = "CDO";
+
+function CDCToken() { return this; }
+CDCToken.prototype = new CSSParserToken;
+CDCToken.prototype.tokenType = "CDC";
+
+function ColonToken() { return this; }
+ColonToken.prototype = new CSSParserToken;
+ColonToken.prototype.tokenType = ":";
+
+function SemicolonToken() { return this; }
+SemicolonToken.prototype = new CSSParserToken;
+SemicolonToken.prototype.tokenType = ";";
+
+function OpenCurlyToken() { return this; }
+OpenCurlyToken.prototype = new CSSParserToken;
+OpenCurlyToken.prototype.tokenType = "{";
+
+function CloseCurlyToken() { return this; }
+CloseCurlyToken.prototype = new CSSParserToken;
+CloseCurlyToken.prototype.tokenType = "}";
+
+function OpenSquareToken() { return this; }
+OpenSquareToken.prototype = new CSSParserToken;
+OpenSquareToken.prototype.tokenType = "[";
+
+function CloseSquareToken() { return this; }
+CloseSquareToken.prototype = new CSSParserToken;
+CloseSquareToken.prototype.tokenType = "]";
+
+function OpenParenToken() { return this; }
+OpenParenToken.prototype = new CSSParserToken;
+OpenParenToken.prototype.tokenType = "(";
+
+function CloseParenToken() { return this; }
+CloseParenToken.prototype = new CSSParserToken;
+CloseParenToken.prototype.tokenType = ")";
+
+function EOFToken() { return this; }
+EOFToken.prototype = new CSSParserToken;
+EOFToken.prototype.tokenType = "EOF";
+
+function DelimToken(code) {
+	this.value = String.fromCharCode(code);
+	return this;
+}
+DelimToken.prototype = new CSSParserToken;
+DelimToken.prototype.tokenType = "DELIM";
+DelimToken.prototype.toString = function() { return "DELIM("+this.value+")"; }
+
+function StringValuedToken() { return this; }
+StringValuedToken.prototype = new CSSParserToken;
+StringValuedToken.prototype.append = function(val) {
+	if(val instanceof Array) {
+		for(var i = 0; i < val.length; i++) {
+			this.value.push(val[i]);
+		}
+	} else {
+		this.value.push(val);
+	}
+	return true;
+}
+StringValuedToken.prototype.finish = function() {
+	this.value = this.valueAsString();
+	return this;
+}
+StringValuedToken.prototype.ASCIImatch = function(str) {
+	return this.valueAsString().toLowerCase() == str.toLowerCase();
+}
+StringValuedToken.prototype.valueAsString = function() {
+	if(typeof this.value == 'string') return this.value;
+	return stringFromCodeArray(this.value);
+}
+StringValuedToken.prototype.valueAsCodes = function() {
+	if(typeof this.value == 'string') {
+		var ret = [];
+		for(var i = 0; i < this.value.length; i++)
+			ret.push(this.value.charCodeAt(i));
+		return ret;
+	}
+	return this.value.filter(function(e){return e;});
+}
+
+function IdentifierToken(val) {
+	this.value = [];
+	this.append(val);
+}
+IdentifierToken.prototype = new StringValuedToken;
+IdentifierToken.prototype.tokenType = "IDENT";
+IdentifierToken.prototype.toString = function() { return "IDENT("+this.value+")"; }
+
+function FunctionToken(val) {
+	// These are always constructed by passing an IdentifierToken
+	this.value = val.finish().value;
+}
+FunctionToken.prototype = new StringValuedToken;
+FunctionToken.prototype.tokenType = "FUNCTION";
+FunctionToken.prototype.toString = function() { return "FUNCTION("+this.value+")"; }
+
+function AtKeywordToken(val) {
+	this.value = [];
+	this.append(val);
+}
+AtKeywordToken.prototype = new StringValuedToken;
+AtKeywordToken.prototype.tokenType = "AT-KEYWORD";
+AtKeywordToken.prototype.toString = function() { return "AT("+this.value+")"; }
+
+function HashToken(val) {
+	this.value = [];
+	this.append(val);
+}
+HashToken.prototype = new StringValuedToken;
+HashToken.prototype.tokenType = "HASH";
+HashToken.prototype.toString = function() { return "HASH("+this.value+")"; }
+
+function StringToken(val) {
+	this.value = [];
+	this.append(val);
+}
+StringToken.prototype = new StringValuedToken;
+StringToken.prototype.tokenType = "STRING";
+StringToken.prototype.toString = function() { return "\""+this.value+"\""; }
+
+function URLToken(val) {
+	this.value = [];
+	this.append(val);
+}
+URLToken.prototype = new StringValuedToken;
+URLToken.prototype.tokenType = "URL";
+URLToken.prototype.toString = function() { return "URL("+this.value+")"; }
+
+function NumberToken(val) {
+	this.value = [];
+	this.append(val);
+	this.type = "integer";
+}
+NumberToken.prototype = new StringValuedToken;
+NumberToken.prototype.tokenType = "NUMBER";
+NumberToken.prototype.toString = function() {
+	if(this.type == "integer")
+		return "INT("+this.value+")";
+	return "NUMBER("+this.value+")";
+}
+NumberToken.prototype.finish = function() {
+	this.repr = this.valueAsString();
+	this.value = this.repr * 1;
+	if(Math.abs(this.value) % 1 != 0) this.type = "number";
+	return this;
+}
+
+function PercentageToken(val) {
+	// These are always created by passing a NumberToken as val
+	val.finish();
+	this.value = val.value;
+	this.repr = val.repr;
+}
+PercentageToken.prototype = new CSSParserToken;
+PercentageToken.prototype.tokenType = "PERCENTAGE";
+PercentageToken.prototype.toString = function() { return "PERCENTAGE("+this.value+")"; }
+
+function DimensionToken(val,unit) {
+	// These are always created by passing a NumberToken as the val
+	val.finish();
+	this.num = val.value;
+	this.unit = [];
+	this.repr = val.repr;
+	this.append(unit);
+}
+DimensionToken.prototype = new CSSParserToken;
+DimensionToken.prototype.tokenType = "DIMENSION";
+DimensionToken.prototype.toString = function() { return "DIM("+this.num+","+this.unit+")"; }
+DimensionToken.prototype.append = function(val) {
+	if(val instanceof Array) {
+		for(var i = 0; i < val.length; i++) {
+			this.unit.push(val[i]);
+		}
+	} else {
+		this.unit.push(val);
+	}
+	return true;
+}
+DimensionToken.prototype.finish = function() {
+	this.unit = stringFromCodeArray(this.unit);
+	this.repr += this.unit;
+	return this;
+}
+
+function UnicodeRangeToken(start,end) {
+	// start and end are array of char codes, completely finished
+	start = parseInt(stringFromCodeArray(start),16);
+	if(end === undefined) end = start + 1;
+	else end = parseInt(stringFromCodeArray(end),16);
+
+	if(start > maximumallowedcodepoint) end = start;
+	if(end < start) end = start;
+	if(end > maximumallowedcodepoint) end = maximumallowedcodepoint;
+
+	this.start = start;
+	this.end = end;
+	return this;
+}
+UnicodeRangeToken.prototype = new CSSParserToken;
+UnicodeRangeToken.prototype.tokenType = "UNICODE-RANGE";
+UnicodeRangeToken.prototype.toString = function() {
+	if(this.start+1 == this.end)
+		return "UNICODE-RANGE("+this.start.toString(16).toUpperCase()+")";
+	if(this.start < this.end)
+		return "UNICODE-RANGE("+this.start.toString(16).toUpperCase()+"-"+this.end.toString(16).toUpperCase()+")";
+	return "UNICODE-RANGE()";
+}
+UnicodeRangeToken.prototype.contains = function(code) {
+	return code >= this.start && code < this.end;
+}
+
+
+// Exportation.
+// TODO: also export the various tokens objects?
+exports.tokenize = tokenize;
+exports.EOFToken = EOFToken;
+
+}));
+
+(function (root, factory) {
+    // Universal Module Definition (UMD) to support AMD, CommonJS/Node.js,
+    // Rhino, and plain browser loading.
+    if (typeof define === 'function' && define.amd) {
+        define('bleach/css-parser/parser',['require', 'exports'], factory);
+    } else if (typeof exports !== 'undefined') {
+        factory(require, exports);
+    } else {
+        factory(root);
+    }
+}(this, function (require, exports) {
+var tokenizer = require('./tokenizer');
+
+function parse(tokens, initialMode) {
+	var mode = initialMode || 'top-level';
+	var i = -1;
+	var token;
+
+	var stylesheet;
+        switch (mode) {
+          case 'top-level':
+             stylesheet = new Stylesheet;
+             break;
+          // (Used for style attributes; start out parsing declarations which
+          // means that our container must be a StyleRule.)
+          case 'declaration':
+             stylesheet = new StyleRule;
+             break;
+        }
+        stylesheet.startTok = tokens[0];
+        //console.log('  initial tok', JSON.stringify(tokens[0]), JSON.stringify(stylesheet));
+	var stack = [stylesheet];
+	var rule = stack[0];
+
+	var consume = function(advance) {
+		if(advance === undefined) advance = 1;
+		i += advance;
+		if(i < tokens.length)
+			token = tokens[i];
+		else
+			token = new EOFToken;
+		return true;
+	};
+	var reprocess = function() {
+		i--;
+		return true;
+	}
+	var next = function() {
+		return tokens[i+1];
+	};
+	var switchto = function(newmode) {
+		if(newmode === undefined) {
+			if(rule.fillType !== '')
+				mode = rule.fillType;
+			else if(rule.type == 'STYLESHEET')
+				mode = 'top-level'
+			else { console.log("Unknown rule-type while switching to current rule's content mode: ",rule); mode = ''; }
+		} else {
+			mode = newmode;
+		}
+		return true;
+	}
+	var push = function(newRule) {
+		rule = newRule;
+                rule.startTok = token;
+                //console.log('  startTok', JSON.stringify(token), JSON.stringify(rule));
+		stack.push(rule);
+		return true;
+	}
+	var parseerror = function(msg) {
+		console.log("Parse error at token " + i + ": " + token + ".\n" + msg);
+		return true;
+	}
+	var pop = function() {
+		var oldrule = stack.pop();
+                oldrule.endTok = token;
+                //console.log('  endTok', JSON.stringify(token), JSON.stringify(oldrule));
+		rule = stack[stack.length - 1];
+		rule.append(oldrule);
+		return true;
+	}
+	var discard = function() {
+		stack.pop();
+		rule = stack[stack.length - 1];
+		return true;
+	}
+	var finish = function() {
+		while(stack.length > 1) {
+			pop();
+		}
+                rule.endTok = token;
+                //console.log('  endTok', JSON.stringify(token), JSON.stringify(rule));
+	}
+
+	for(;;) {
+		consume();
+
+		switch(mode) {
+		case "top-level":
+			switch(token.tokenType) {
+			case "CDO":
+			case "CDC":
+			case "WHITESPACE": break;
+			case "AT-KEYWORD": push(new AtRule(token.value)) && switchto('at-rule'); break;
+			case "{": parseerror("Attempt to open a curly-block at top-level.") && consumeAPrimitive(); break;
+			case "EOF": finish(); return stylesheet;
+			default: push(new StyleRule) && switchto('selector') && reprocess();
+			}
+			break;
+
+		case "at-rule":
+			switch(token.tokenType) {
+			case ";": pop() && switchto(); break;
+			case "{":
+				if(rule.fillType !== '') switchto(rule.fillType);
+				else parseerror("Attempt to open a curly-block in a statement-type at-rule.") && discard() && switchto('next-block') && reprocess();
+				break;
+			case "EOF": finish(); return stylesheet;
+			default: rule.appendPrelude(consumeAPrimitive());
+			}
+			break;
+
+		case "rule":
+			switch(token.tokenType) {
+			case "WHITESPACE": break;
+			case "}": pop() && switchto(); break;
+			case "AT-KEYWORD": push(new AtRule(token.value)) && switchto('at-rule'); break;
+			case "EOF": finish(); return stylesheet;
+			default: push(new StyleRule) && switchto('selector') && reprocess();
+			}
+			break;
+
+		case "selector":
+			switch(token.tokenType) {
+			case "{": switchto('declaration'); break;
+			case "EOF": discard() && finish(); return stylesheet;
+			default: rule.appendSelector(consumeAPrimitive());
+			}
+			break;
+
+		case "declaration":
+			switch(token.tokenType) {
+			case "WHITESPACE":
+			case ";": break;
+			case "}": pop() && switchto(); break;
+			case "AT-RULE": push(new AtRule(token.value)) && switchto('at-rule'); break;
+			case "IDENT": push(new Declaration(token.value)) && switchto('after-declaration-name'); break;
+			case "EOF": finish(); return stylesheet;
+			default: parseerror() && discard() && switchto('next-declaration');
+			}
+			break;
+
+		case "after-declaration-name":
+			switch(token.tokenType) {
+			case "WHITESPACE": break;
+			case ":": switchto('declaration-value'); break;
+			case ";": parseerror("Incomplete declaration - semicolon after property name.") && discard() && switchto(); break;
+			case "EOF": discard() && finish(); return stylesheet;
+			default: parseerror("Invalid declaration - additional token after property name") && discard() && switchto('next-declaration');
+			}
+			break;
+
+		case "declaration-value":
+			switch(token.tokenType) {
+			case "DELIM":
+				if(token.value == "!" && next().tokenType == 'IDENTIFIER' && next().value.toLowerCase() == "important") {
+					consume();
+					rule.important = true;
+					switchto('declaration-end');
+				} else {
+					rule.append(token);
+				}
+				break;
+			case ";": pop() && switchto(); break;
+			case "}": pop() && pop() && switchto(); break;
+			case "EOF": finish(); return stylesheet;
+			default: rule.append(consumeAPrimitive());
+			}
+			break;
+
+		case "declaration-end":
+			switch(token.tokenType) {
+			case "WHITESPACE": break;
+			case ";": pop() && switchto(); break;
+			case "}": pop() && pop() && switchto(); break;
+			case "EOF": finish(); return stylesheet;
+			default: parseerror("Invalid declaration - additional token after !important.") && discard() && switchto('next-declaration');
+			}
+			break;
+
+		case "next-block":
+			switch(token.tokenType) {
+			case "{": consumeAPrimitive() && switchto(); break;
+			case "EOF": finish(); return stylesheet;
+			default: consumeAPrimitive(); break;
+			}
+			break;
+
+		case "next-declaration":
+			switch(token.tokenType) {
+			case ";": switchto('declaration'); break;
+			case "}": switchto('declaration') && reprocess(); break;
+			case "EOF": finish(); return stylesheet;
+			default: consumeAPrimitive(); break;
+			}
+			break;
+
+		default:
+			// If you hit this, it's because one of the switchto() calls is typo'd.
+			console.log('Unknown parsing mode: ' + mode);
+			return;
+		}
+	}
+
+	function consumeAPrimitive() {
+		switch(token.tokenType) {
+		case "(":
+		case "[":
+		case "{": return consumeASimpleBlock();
+		case "FUNCTION": return consumeAFunc();
+		default: return token;
+		}
+	}
+
+	function consumeASimpleBlock() {
+		var endingTokenType = {"(":")", "[":"]", "{":"}"}[token.tokenType];
+		var block = new SimpleBlock(token.tokenType);
+
+		for(;;) {
+			consume();
+			switch(token.tokenType) {
+			case "EOF":
+			case endingTokenType: return block;
+			default: block.append(consumeAPrimitive());
+			}
+		}
+	}
+
+	function consumeAFunc() {
+		var func = new Func(token.value);
+		var arg = new FuncArg();
+
+		for(;;) {
+			consume();
+			switch(token.tokenType) {
+			case "EOF":
+			case ")": func.append(arg); return func;
+			case "DELIM":
+				if(token.value == ",") {
+					func.append(arg);
+					arg = new FuncArg();
+				} else {
+					arg.append(token);
+				}
+				break;
+			default: arg.append(consumeAPrimitive());
+			}
+		}
+	}
+}
+
+function CSSParserRule() { return this; }
+CSSParserRule.prototype.fillType = '';
+CSSParserRule.prototype.toString = function(indent) {
+	return JSON.stringify(this.toJSON(),null,indent);
+}
+CSSParserRule.prototype.append = function(val) {
+	this.value.push(val);
+	return this;
+}
+
+function Stylesheet() {
+	this.value = [];
+	return this;
+}
+Stylesheet.prototype = new CSSParserRule;
+Stylesheet.prototype.type = "STYLESHEET";
+Stylesheet.prototype.toJSON = function() {
+	return {type:'stylesheet', value: this.value.map(function(e){return e.toJSON();})};
+}
+
+function AtRule(name) {
+	this.name = name;
+	this.prelude = [];
+	this.value = [];
+	if(name in AtRule.registry)
+		this.fillType = AtRule.registry[name];
+	return this;
+}
+AtRule.prototype = new CSSParserRule;
+AtRule.prototype.type = "AT-RULE";
+AtRule.prototype.appendPrelude = function(val) {
+	this.prelude.push(val);
+	return this;
+}
+AtRule.prototype.toJSON = function() {
+	return {type:'at', name:this.name, prelude:this.prelude.map(function(e){return e.toJSON();}), value:this.value.map(function(e){return e.toJSON();})};
+}
+AtRule.registry = {
+	'import': '',
+	'media': 'rule',
+	'font-face': 'declaration',
+	'page': 'declaration',
+	'keyframes': 'rule',
+	'namespace': '',
+	'counter-style': 'declaration',
+	'supports': 'rule',
+	'document': 'rule',
+	'font-feature-values': 'declaration',
+	'viewport': '',
+	'region-style': 'rule'
+};
+
+function StyleRule() {
+	this.selector = [];
+	this.value = [];
+	return this;
+}
+StyleRule.prototype = new CSSParserRule;
+StyleRule.prototype.type = "STYLE-RULE";
+StyleRule.prototype.fillType = 'declaration';
+StyleRule.prototype.appendSelector = function(val) {
+	this.selector.push(val);
+	return this;
+}
+StyleRule.prototype.toJSON = function() {
+	return {type:'selector', selector:this.selector.map(function(e){return e.toJSON();}), value:this.value.map(function(e){return e.toJSON();})};
+}
+
+function Declaration(name) {
+	this.name = name;
+	this.value = [];
+	return this;
+}
+Declaration.prototype = new CSSParserRule;
+Declaration.prototype.type = "DECLARATION";
+Declaration.prototype.toJSON = function() {
+	return {type:'declaration', name:this.name, value:this.value.map(function(e){return e.toJSON();})};
+}
+
+function SimpleBlock(type) {
+	this.name = type;
+	this.value = [];
+	return this;
+}
+SimpleBlock.prototype = new CSSParserRule;
+SimpleBlock.prototype.type = "BLOCK";
+SimpleBlock.prototype.toJSON = function() {
+	return {type:'block', name:this.name, value:this.value.map(function(e){return e.toJSON();})};
+}
+
+function Func(name) {
+	this.name = name;
+	this.value = [];
+	return this;
+}
+Func.prototype = new CSSParserRule;
+Func.prototype.type = "FUNCTION";
+Func.prototype.toJSON = function() {
+	return {type:'func', name:this.name, value:this.value.map(function(e){return e.toJSON();})};
+}
+
+function FuncArg() {
+	this.value = [];
+	return this;
+}
+FuncArg.prototype = new CSSParserRule;
+FuncArg.prototype.type = "FUNCTION-ARG";
+FuncArg.prototype.toJSON = function() {
+	return this.value.map(function(e){return e.toJSON();});
+}
+
+// Exportation.
+// TODO: also export the various rule objects?
+exports.parse = parse;
+
+}));
+
+if (typeof exports === 'object' && typeof define !== 'function') {
+    define = function (factory) {
+        factory(require, exports, module);
+    };
+}
+
+define('bleach',['require','exports','module','./bleach/css-parser/tokenizer','./bleach/css-parser/parser'],function (require, exports, module) {
+var tokenizer = require('./bleach/css-parser/tokenizer');
+var parser = require('./bleach/css-parser/parser');
+
+var ALLOWED_TAGS = [
+    'a',
+    'abbr',
+    'acronym',
+    'b',
+    'blockquote',
+    'code',
+    'em',
+    'i',
+    'li',
+    'ol',
+    'strong',
+    'ul'
+];
+var ALLOWED_ATTRIBUTES = {
+    'a': ['href', 'title'],
+    'abbr': ['title'],
+    'acronym': ['title']
+};
+var ALLOWED_STYLES = [];
+
+var Node = {
+  ELEMENT_NODE                :  1,
+  ATTRIBUTE_NODE              :  2,
+  TEXT_NODE                   :  3,
+  CDATA_SECTION_NODE          :  4,
+  ENTITY_REFERENCE_NODE       :  5,
+  ENTITY_NODE                 :  6,
+  PROCESSING_INSTRUCTION_NODE :  7,
+  COMMENT_NODE                :  8,
+  DOCUMENT_NODE               :  9,
+  DOCUMENT_TYPE_NODE          : 10,
+  DOCUMENT_FRAGMENT_NODE      : 11,
+  NOTATION_NODE               : 12
+};
+
+var DEFAULTS = {
+  tags: ALLOWED_TAGS,
+  prune: [],
+  attributes: ALLOWED_ATTRIBUTES,
+  styles: ALLOWED_STYLES,
+  strip: false,
+  stripComments: true
+};
+
+/**
+ * Clean a string.
+ */
+exports.clean = function (html, opts) {
+  if (!html) return '';
+
+  // This is poor's man doctype/meta cleanup. I wish DOMParser works in a
+  // worker but it sounds like a dream, see bug 677123.
+  // Someone needs to come with a better approach but I'm running out of
+  // time...
+  // Prevoiusly, only removed DOCTYPE at start of string, but some HTML
+  // senders are so bad they just leave them in the middle of email
+  // content, as if they just dump from their CMS. So removing all of them
+  // now
+  html = html.replace(/<!DOCTYPE\s+[^>]*>/g, '');
+
+  return exports.cleanNode(html, opts);
+};
+
+
+/**
+ */
+exports.cleanNode = function(html, opts) {
+try {
+  function debug(str) {
+    console.log("Bleach: " + str + "\n");
+  }
+
+  opts = opts || DEFAULTS;
+
+  var attrsByTag = opts.hasOwnProperty('attributes') ?
+                    opts.attributes : DEFAULTS.attributes;
+  var wildAttrs;
+  if (Array.isArray(attrsByTag)) {
+    wildAttrs = attrsByTag;
+    attrsByTag = {};
+  } else if (attrsByTag.hasOwnProperty('*')) {
+    wildAttrs = attrsByTag['*'];
+  } else {
+    wildAttrs = [];
+  }
+  var sanitizeOptions = {
+    ignoreComment: ('stripComments' in opts) ? opts.stripComments
+                                             : DEFAULTS.stripComments,
+    allowedStyles: opts.styles || DEFAULTS.styles,
+    allowedTags: opts.tags || DEFAULTS.tags,
+    stripMode: ('strip' in opts) ? opts.strip : DEFAULTS.strip,
+    pruneTags: opts.prune || DEFAULTS.prune,
+    allowedAttributesByTag: attrsByTag,
+    wildAttributes: wildAttrs,
+    callbackRegexp: opts.callbackRegexp || null,
+    callback: opts.callbackRegexp && opts.callback || null,
+    maxLength: opts.maxLength || 0
+  };
+
+  var sanitizer = new HTMLSanitizer(sanitizeOptions);
+  HTMLParser.HTMLParser(html, sanitizer);
+  return sanitizer.output;
+} catch(e) {
+  console.error(e, '\n', e.stack);
+  throw e;
+}
+
+};
+
+var RE_NORMALIZE_WHITESPACE = /\s+/g;
+
+var HTMLSanitizer = function(options) {
+  this.output = '';
+
+  this.ignoreComment = options.ignoreComment;
+  this.allowedStyles = options.allowedStyles;
+  this.allowedTags = options.allowedTags;
+  this.stripMode = options.stripMode;
+  this.pruneTags = options.pruneTags;
+  this.allowedAttributesByTag = options.allowedAttributesByTag;
+  this.wildAttributes = options.wildAttributes;
+
+  this.callbackRegexp = options.callbackRegexp;
+  this.callback = options.callback;
+
+  this.isInsideStyleTag = false;
+  // How many pruned tag types are on the stack; we require them to be fully
+  // balanced, but don't care if what's inside them is balanced or not.
+  this.isInsidePrunedTag = 0;
+  // Similar; not clear why we need to bother counting for these. debug?
+  this.isInsideStrippedTag = 0;
+
+  // Added to allow snippet generation. Pass in
+  // maxLength to get snippet work.
+  this.maxLength = options.maxLength || 0;
+
+  // Flag to indicate parsing should not
+  // continue because maxLength has been hit.
+  this.complete = false;
+
+  // If just getting a snippet, the input
+  // may also just be an HTML snippet, so
+  // if parsing cannot continue, signal
+  // just to stop at that point.
+  this.ignoreFragments = this.maxLength > 0;
+};
+
+HTMLSanitizer.prototype = {
+  start: function(tag, attrs, unary) {
+    // - prune (trumps all else)
+    if (this.pruneTags.indexOf(tag) !== -1) {
+      if (!unary)
+        this.isInsidePrunedTag++;
+      return;
+    }
+    else if (this.isInsidePrunedTag) {
+      return;
+    }
+    // - strip
+    if (this.allowedTags.indexOf(tag) === -1) {
+      // In strip mode we discard the tag rather than escaping it.
+      if (this.stripMode) {
+        if (!unary) {
+          this.isInsideStrippedTag++;
+        }
+        return;
+      }
+
+      // The tag is not in the whitelist
+      this.output += "&lt;" + (unary ? "/" : "") + tag + "&gt;";
+      return;
+    }
+
+    this.isInsideStyleTag = (tag == "style" && !unary);
+
+    // If a callback was specified and it matches the tag name, then invoke
+    // the callback.  This happens before the attribute filtering so that
+    // the function can observe dangerous attributes, but in the event of
+    // the (silent) failure of this function, they will still be safely
+    // removed.
+    var callbackRegexp = this.callbackRegexp;
+    if (callbackRegexp && callbackRegexp.test(tag)) {
+      attrs = this.callback(tag, attrs);
+    }
+
+    var whitelist = this.allowedAttributesByTag[tag];
+    var wildAttrs = this.wildAttributes;
+    var result = "<" + tag;
+    for (var i = 0; i < attrs.length; i++) {
+      var attr = attrs[i];
+      var attrName = attr.name.toLowerCase();
+
+      if (wildAttrs.indexOf(attrName) !== -1 ||
+          (whitelist && whitelist.indexOf(attrName) !== -1)) {
+        if (attrName == "style") {
+          var attrValue = '';
+          try {
+            attrValue = CSSParser.parseAttribute(attr.escaped,
+                                                   this.allowedStyles);
+          } catch (e) {
+            console.log('CSSParser.parseAttribute failed for: "' +
+                         attr.escaped + '", skipping. Error: ' + e);
+          }
+          result += " " + attrName + '="' + attrValue + '"';
+        } else {
+          result += " " + attrName + '="' + attr.escaped + '"';
+        }
+      }
+    }
+    result += (unary ? "/" : "") + ">";
+
+    this.output += result;
+  },
+
+  end: function(tag) {
+    if (this.pruneTags.indexOf(tag) !== -1) {
+      this.isInsidePrunedTag--;
+      return;
+    }
+    else if (this.isInsidePrunedTag) {
+      return;
+    }
+
+    if (this.allowedTags.indexOf(tag) === -1) {
+      if (this.isInsideStrippedTag) {
+        this.isInsideStrippedTag--;
+        return;
+      }
+
+      this.output += "&lt;/" + tag + "&gt;";
+      return;
+    }
+
+    if (this.isInsideStyleTag) {
+      this.isInsideStyleTag = false;
+    }
+
+    this.output += "</" + tag + ">";
+  },
+
+  chars: function(text) {
+    if (this.isInsidePrunedTag || this.complete)
+      return;
+    if (this.isInsideStyleTag) {
+      this.output += CSSParser.parseBody(text, this.allowedStyles);
+      return;
+    }
+
+    //console.log('HTML SANITIZER CHARS GIVEN: ' + text);
+    if (this.maxLength) {
+      if (this.insideTagForSnippet) {
+        if (text.indexOf('>') !== -1) {
+          // All clear now, for the next chars call
+          this.insideTagForSnippet = false;
+        }
+        return;
+      } else {
+        // Skip chars that are for a tag, not wanted for a snippet.
+        if (text.charAt(0) === '<') {
+          this.insideTagForSnippet = true;
+          return;
+        }
+      }
+
+      // the whitespace down to one whitespace character.
+      var normalizedText = text.replace(RE_NORMALIZE_WHITESPACE, ' ');
+
+      // If the join would create two adjacents spaces, then skip the one
+      // on the thing we are concatenating.
+      var length = this.output.length;
+      if (length && normalizedText[0] === ' ' &&
+          this.output[length - 1] === ' ') {
+        normalizedText = normalizedText.substring(1);
+      }
+
+      this.output += normalizedText;
+      if (this.output.length >= this.maxLength) {
+        this.output = this.output.substring(0, this.maxLength);
+        // XXX We got the right numbers of chars
+        // Do not process anymore, and also set state
+        // the parser can use to know to stop doing work.
+        this.complete = true;
+      }
+    } else {
+      this.output += escapeHTMLTextKeepingExistingEntities(text);
+    }
+  },
+
+  comment: function(comment) {
+    if (this.isInsidePrunedTag)
+      return;
+    if (this.ignoreComment)
+      return;
+    this.output += '<!--' + comment + '-->';
+  }
+};
+
+/*
+ * HTML Parser By John Resig (ejohn.org)
+ * Although the file only calls out MPL as a valid license, the upstream is
+ * available under Apache 2.0 and John Resig has indicated by e-mail to
+ * asuth@mozilla.com on 2013-03-13 that Apache 2.0 is fine.  So we are using
+ * it under Apache 2.0.
+ * http://ejohn.org/blog/pure-javascript-html-parser/
+ *
+ * Original code by Erik Arvidsson, tri-licensed under Apache 2.0, MPL 1.1
+ * (probably implicitly 1.1+), or GPL 2.0+ (as visible in the file):
+ * http://erik.eae.net/simplehtmlparser/simplehtmlparser.js
+ *
+ * // Use like so:
+ * HTMLParser(htmlString, {
+ *     start: function(tag, attrs, unary) {},
+ *     end: function(tag) {},
+ *     chars: function(text) {},
+ *     comment: function(text) {}
+ * });
+ *
+ */
+
+
+var HTMLParser = (function(){
+  // Important syntax notes from the WHATWG HTML spec and observations.
+  // http://www.whatwg.org/specs/web-apps/current-work/multipage/syntax.html
+  // http://www.whatwg.org/specs/web-apps/current-work/multipage/common-microsyntaxes.html#common-parser-idioms
+  //
+  // The spec says _html_ tag names are [A-Za-z0-9]; we also include '-' and '_'
+  // because that's what the code already did, but also since Gecko seems to be
+  // very happy to parse those characters.
+  //
+  // The spec defines attributes by what they must not include, which is:
+  // [\0\s"'>/=] plus also no control characters, or non-unicode characters.
+  //
+  // The (inherited) code used to have the regular expression effectively
+  // validate the attribute syntax by including their grammer in the regexp.
+  // The problem with this is that it can make the regexp fail to match tags
+  // that are clearly tags.  When we encountered (quoted) attributes without
+  // whitespace between them, we would escape the entire tag.  Attempted
+  // trivial fixes resulted in regex back-tracking, which begged the issue of
+  // why the regex would do this in the first place.  So we stopped doing that.
+  //
+  // CDATA *is not a thing* in the HTML namespace.  <![CDATA[ just gets treated
+  // as a "bogus comment".  See:
+  // http://www.whatwg.org/specs/web-apps/current-work/multipage/tokenization.html#markup-declaration-open-state
+
+  // NOTE: tag and attr regexps changed to ignore name spaces prefixes!
+  //
+  // CHANGE: "we" previously required there to be white-space between attributes.
+  // Unfortunately, the world does not agree with this, so we now require
+  // whitespace only after the tag name prior to the first attribute and make
+  // the whole attribute clause optional.
+  //
+  // - Regular Expressions for parsing tags and attributes
+  // ^<                     anchored tag open character
+  // (?:[-A-Za-z0-9_]+:)?   eat the namespace
+  // ([-A-Za-z0-9_]+)       the tag name
+  // ([^>]*)                capture attributes and/or closing '/' if present
+  // >                      tag close character
+  var startTag = /^<(?:[-A-Za-z0-9_]+:)?([-A-Za-z0-9_]+)([^>]*)>/,
+  // ^<\/                   close tag lead-in
+  // (?:[-A-Za-z0-9_]+:)?   optional tag prefix
+  // ([-A-Za-z0-9_]+)       tag name
+  // [^>]*                  The spec says this should be whitespace, we forgive.
+  // >
+    endTag = /^<\/(?:[-A-Za-z0-9_]+:)?([-A-Za-z0-9_]+)[^>]*>/,
+  // NOTE: This regexp was doing something freaky with the value quotings
+  // before. (?:"((?:\\.|[^"])*)") instead of (?:"[^"]*") from the tag part,
+  // which is deeply confusing.  Since the period thing seems meaningless, I am
+  // replacing it from the bits from startTag
+  //
+  // (?:[-A-Za-z0-9_]+:)?   attribute prefix
+  // ([-A-Za-z0-9_]+)       attribute name
+  // (?:                    The attribute doesn't need a value
+  //  \s*=\s*               whitespace, = to indicate value, whitespace
+  //  (?:                   attribute values:
+  //   (?:"([^"]*)")|       capture double-quoted
+  //   (?:'([^']*)')|       capture single-quoted
+  //   ([^>\s]+)            capture unquoted
+  //  )
+  // )?                    (the attribute does't need a value)
+    attr = /(?:[-A-Za-z0-9_]+:)?([-A-Za-z0-9_]+)(?:\s*=\s*(?:(?:"([^"]*)")|(?:'([^']*)')|([^>\s]+)))?/g;
+
+  // - Empty Elements - HTML 4.01
+  var empty = makeMap("area,base,basefont,br,col,frame,hr,img,input,isindex,link,meta,param,embed");
+
+  // - Block Elements - HTML 4.01
+  var block = makeMap("address,applet,blockquote,button,center,dd,del,dir,div,dl,dt,fieldset,form,frameset,hr,iframe,ins,isindex,li,map,menu,noframes,noscript,object,ol,p,pre,script,table,tbody,td,tfoot,th,thead,tr,ul");
+
+  // - Inline Elements - HTML 4.01
+  var inline = makeMap("a,abbr,acronym,applet,b,basefont,bdo,big,br,button,cite,code,del,dfn,em,font,i,iframe,img,input,ins,kbd,label,map,object,q,s,samp,script,select,small,span,strike,strong,sub,sup,textarea,tt,u,var");
+
+  // - Elements that you can, intentionally, leave open (and close themselves)
+  var closeSelf = makeMap("colgroup,dd,dt,li,options,p,td,tfoot,th,thead,tr");
+
+  // - Attributes that have their values filled in disabled="disabled"
+  var fillAttrs = makeMap("checked,compact,declare,defer,disabled,ismap,multiple,nohref,noresize,noshade,nowrap,readonly,selected");
+
+  // - Special Elements (can contain anything)
+  var special = makeMap("script,style");
+
+  var HTMLParser = this.HTMLParser = function( html, handler ) {
+    var index, chars, match, stack = [], last = html;
+    stack.last = function(){
+      return this[ this.length - 1 ];
+    };
+
+    while ( html ) {
+      chars = true;
+
+      // Make sure we're not in a script or style element
+      if ( !stack.last() || !special[ stack.last() ] ) {
+
+        // Comment
+        if ( html.lastIndexOf("<!--", 0) == 0 ) {
+          index = html.indexOf("-->");
+
+                                        // WHATWG spec says the text can't start
+                                        // with the closing tag.
+          if ( index >= 5 ) {
+            if ( handler.comment )
+              handler.comment( html.substring( 4, index ) );
+            html = html.substring( index + 3 );
+            chars = false;
+          } else {
+            // The comment does not have a end. Let's return the whole string as a comment then.
+            if ( handler.comment )
+              handler.comment( html.substring( 4, -1 ) );
+            html = '';
+            chars = false;
+          }
+
+        // end tag
+        } else if ( html.lastIndexOf("</", 0) == 0 ) {
+          match = html.match( endTag );
+
+          if ( match ) {
+            html = html.substring( match[0].length );
+            match[0].replace( endTag, parseEndTag );
+            chars = false;
+          }
+
+        // start tag
+        } else if ( html.lastIndexOf("<", 0) == 0 ) {
+          match = html.match( startTag );
+
+          if ( match ) {
+            html = html.substring( match[0].length );
+            match[0].replace( startTag, parseStartTag );
+            chars = false;
+          }
+        }
+
+        if ( chars ) {
+          index = html.indexOf("<");
+
+          if (index === 0) {
+            // This is not a valid tag in regards of the parser.
+            var text = html.substring(0, 1);
+            html = html.substring(1);
+          } else {
+            var text = index < 0 ? html : html.substring( 0, index );
+            html = index < 0 ? "" : html.substring( index );
+          }
+
+          if ( handler.chars ) {
+            handler.chars( text );
+            if ( handler.complete )
+              return this;
+          }
+        }
+
+      } else { // specials: script or style
+        var skipWork = false;
+        html = html.replace(
+          // we use "[^]" instead of "." because it matches newlines too
+          new RegExp("^([^]*?)<\/" + stack.last() + "[^>]*>", "i"),
+          function(all, text){
+            if (!skipWork) {
+              text = text.replace(/<!--([^]*?)-->/g, "$1")
+                .replace(/<!\[CDATA\[([^]*?)]]>/g, "$1");
+
+              if ( handler.chars ) {
+                handler.chars( text );
+                skipWork = handler.complete;
+              }
+            }
+
+            return "";
+          });
+
+        if ( handler.complete )
+          return this;
+
+        parseEndTag( "", stack.last() );
+      }
+
+      if ( html == last ) {
+        // May just have a fragment of HTML, to
+        // generate a snippet. If that is the case
+        // just end parsing now.
+        if ( handler.ignoreFragments ) {
+          return;
+        } else {
+          console.log(html);
+          console.log(last);
+          throw "Parse Error: " + html;
+        }
+      }
+      last = html;
+    }
+
+    // Clean up any remaining tags
+    parseEndTag();
+
+    function parseStartTag( tag, tagName, rest ) {
+      tagName = tagName.toLowerCase();
+      if ( block[ tagName ] ) {
+        while ( stack.last() && inline[ stack.last() ] ) {
+          parseEndTag( "", stack.last() );
+        }
+      }
+
+      if ( closeSelf[ tagName ] && stack.last() == tagName ) {
+        parseEndTag( "", tagName );
+      }
+
+      var unary = empty[ tagName ];
+      // to simplify the regexp, the 'rest capture group now absorbs the /, so
+      // we need to strip it off if it's there.
+      if (rest.length && rest[rest.length - 1] === '/') {
+        unary = true;
+        rest = rest.slice(0, -1);
+      }
+
+      if ( !unary )
+        stack.push( tagName );
+
+      if ( handler.start ) {
+        var attrs = [];
+
+        rest.replace(attr, function(match, name) {
+          // The attr regexp capture groups:
+          // 1: attribute name
+          // 2: double-quoted attribute value (whitespace allowed inside)
+          // 3: single-quoted attribute value (whitespace allowed inside)
+          // 4: un-quoted attribute value (whitespace forbidden)
+          // We need to escape double-quotes because of the risks in there.
+          var value = arguments[2] ? arguments[2] :
+            arguments[3] ? arguments[3] :
+            arguments[4] ? arguments[4] :
+            fillAttrs[name] ? name : "";
+
+          attrs.push({
+            name: name,
+            value: value,
+            escaped: value.replace(/"/g, '&quot;')
+          });
+        });
+
+        if ( handler.start )
+          handler.start( tagName, attrs, unary );
+      }
+    }
+
+    function parseEndTag( tag, tagName ) {
+      // If no tag name is provided, clean shop
+      if ( !tagName )
+        var pos = 0;
+
+      // Find the closest opened tag of the same type
+      else {
+        tagName = tagName.toLowerCase();
+        for ( var pos = stack.length - 1; pos >= 0; pos-- )
+          if ( stack[ pos ] == tagName )
+            break;
+      }
+
+      if ( pos >= 0 ) {
+        // Close all the open elements, up the stack
+        for ( var i = stack.length - 1; i >= pos; i-- )
+          if ( handler.end )
+            handler.end( stack[ i ] );
+
+        // Remove the open elements from the stack
+        stack.length = pos;
+      }
+    }
+  };
+
+  function makeMap(str){
+    var obj = {}, items = str.split(",");
+    for ( var i = 0; i < items.length; i++ )
+      obj[ items[i] ] = true;
+    return obj;
+  }
+
+  return this;
+})();
+
+var CSSParser = {
+  parseAttribute: function (data, allowedStyles) {
+    var tokens = tokenizer.tokenize(data, { loc: true });
+    var rule = parser.parse(tokens, 'declaration');
+
+    var keepText = [];
+    this._filterDeclarations(null, rule.value, allowedStyles, data, keepText);
+    var oot = keepText.join('');
+    //console.log('IN:', data, '\n OUT:', oot);
+    return oot;
+  },
+
+  _filterDeclarations: function(parent, decls, allowedStyles, fullText,
+                                textOut) {
+    for (var i = 0; i < decls.length; i++) {
+      var decl = decls[i];
+      if (decl.type !== 'DECLARATION') {
+        continue;
+      }
+      if (allowedStyles.indexOf(decl.name) !== -1) {
+        textOut.push(fullText.substring(
+          decl.startTok.loc.start.idx,
+          // If we have a parent and our parent ends on the same token as us,
+          // then don't emit our ending token (ex: '}'), otherwise do emit it
+          // (ex: ';').  We don't want a parent when it's synthetic like for
+          // parseAttribute.
+          (parent && parent.endTok === decl.endTok) ?
+            decl.endTok.loc.start.idx :
+            decl.endTok.loc.end.idx + 1));
+      }
+    }
+  },
+
+  parseBody: function (data, allowedStyles) {
+    var body = "";
+    var oot = "";
+
+    try {
+      var tokens = tokenizer.tokenize(data, { loc: true });
+      var stylesheet = parser.parse(tokens);
+
+      var keepText = [];
+
+      for (var i = 0; i < stylesheet.value.length; i++) {
+        var sub = stylesheet.value[i];
+        if (sub.type === 'STYLE-RULE') {
+          // We want our tokens up to the start of our first child.  If we have
+          // no children, just go up to the start of our ending token.
+          keepText.push(data.substring(
+            sub.startTok.loc.start.idx,
+            sub.value.length ? sub.value[0].startTok.loc.start.idx
+                             : sub.endTok.loc.start.idx));
+          this._filterDeclarations(sub, sub.value, allowedStyles, data, keepText);
+          // we want all of our terminating token.
+          keepText.push(data.substring(
+            sub.endTok.loc.start.idx, sub.endTok.loc.end.idx + 1));
+        }
+      }
+
+      oot = keepText.join('');
+    } catch (e) {
+      console.log('bleach CSS parsing failed, skipping. Error: ' + e);
+      oot = '';
+    }
+
+    //console.log('IN:', data, '\n OUT:', oot);
+    return oot;
+  }
+};
+
+
+var entities = {
+  34 : 'quot',
+  38 : 'amp',
+  39 : 'apos',
+  60 : 'lt',
+  62 : 'gt',
+  160 : 'nbsp',
+  161 : 'iexcl',
+  162 : 'cent',
+  163 : 'pound',
+  164 : 'curren',
+  165 : 'yen',
+  166 : 'brvbar',
+  167 : 'sect',
+  168 : 'uml',
+  169 : 'copy',
+  170 : 'ordf',
+  171 : 'laquo',
+  172 : 'not',
+  173 : 'shy',
+  174 : 'reg',
+  175 : 'macr',
+  176 : 'deg',
+  177 : 'plusmn',
+  178 : 'sup2',
+  179 : 'sup3',
+  180 : 'acute',
+  181 : 'micro',
+  182 : 'para',
+  183 : 'middot',
+  184 : 'cedil',
+  185 : 'sup1',
+  186 : 'ordm',
+  187 : 'raquo',
+  188 : 'frac14',
+  189 : 'frac12',
+  190 : 'frac34',
+  191 : 'iquest',
+  192 : 'Agrave',
+  193 : 'Aacute',
+  194 : 'Acirc',
+  195 : 'Atilde',
+  196 : 'Auml',
+  197 : 'Aring',
+  198 : 'AElig',
+  199 : 'Ccedil',
+  200 : 'Egrave',
+  201 : 'Eacute',
+  202 : 'Ecirc',
+  203 : 'Euml',
+  204 : 'Igrave',
+  205 : 'Iacute',
+  206 : 'Icirc',
+  207 : 'Iuml',
+  208 : 'ETH',
+  209 : 'Ntilde',
+  210 : 'Ograve',
+  211 : 'Oacute',
+  212 : 'Ocirc',
+  213 : 'Otilde',
+  214 : 'Ouml',
+  215 : 'times',
+  216 : 'Oslash',
+  217 : 'Ugrave',
+  218 : 'Uacute',
+  219 : 'Ucirc',
+  220 : 'Uuml',
+  221 : 'Yacute',
+  222 : 'THORN',
+  223 : 'szlig',
+  224 : 'agrave',
+  225 : 'aacute',
+  226 : 'acirc',
+  227 : 'atilde',
+  228 : 'auml',
+  229 : 'aring',
+  230 : 'aelig',
+  231 : 'ccedil',
+  232 : 'egrave',
+  233 : 'eacute',
+  234 : 'ecirc',
+  235 : 'euml',
+  236 : 'igrave',
+  237 : 'iacute',
+  238 : 'icirc',
+  239 : 'iuml',
+  240 : 'eth',
+  241 : 'ntilde',
+  242 : 'ograve',
+  243 : 'oacute',
+  244 : 'ocirc',
+  245 : 'otilde',
+  246 : 'ouml',
+  247 : 'divide',
+  248 : 'oslash',
+  249 : 'ugrave',
+  250 : 'uacute',
+  251 : 'ucirc',
+  252 : 'uuml',
+  253 : 'yacute',
+  254 : 'thorn',
+  255 : 'yuml',
+  402 : 'fnof',
+  913 : 'Alpha',
+  914 : 'Beta',
+  915 : 'Gamma',
+  916 : 'Delta',
+  917 : 'Epsilon',
+  918 : 'Zeta',
+  919 : 'Eta',
+  920 : 'Theta',
+  921 : 'Iota',
+  922 : 'Kappa',
+  923 : 'Lambda',
+  924 : 'Mu',
+  925 : 'Nu',
+  926 : 'Xi',
+  927 : 'Omicron',
+  928 : 'Pi',
+  929 : 'Rho',
+  931 : 'Sigma',
+  932 : 'Tau',
+  933 : 'Upsilon',
+  934 : 'Phi',
+  935 : 'Chi',
+  936 : 'Psi',
+  937 : 'Omega',
+  945 : 'alpha',
+  946 : 'beta',
+  947 : 'gamma',
+  948 : 'delta',
+  949 : 'epsilon',
+  950 : 'zeta',
+  951 : 'eta',
+  952 : 'theta',
+  953 : 'iota',
+  954 : 'kappa',
+  955 : 'lambda',
+  956 : 'mu',
+  957 : 'nu',
+  958 : 'xi',
+  959 : 'omicron',
+  960 : 'pi',
+  961 : 'rho',
+  962 : 'sigmaf',
+  963 : 'sigma',
+  964 : 'tau',
+  965 : 'upsilon',
+  966 : 'phi',
+  967 : 'chi',
+  968 : 'psi',
+  969 : 'omega',
+  977 : 'thetasym',
+  978 : 'upsih',
+  982 : 'piv',
+  8226 : 'bull',
+  8230 : 'hellip',
+  8242 : 'prime',
+  8243 : 'Prime',
+  8254 : 'oline',
+  8260 : 'frasl',
+  8472 : 'weierp',
+  8465 : 'image',
+  8476 : 'real',
+  8482 : 'trade',
+  8501 : 'alefsym',
+  8592 : 'larr',
+  8593 : 'uarr',
+  8594 : 'rarr',
+  8595 : 'darr',
+  8596 : 'harr',
+  8629 : 'crarr',
+  8656 : 'lArr',
+  8657 : 'uArr',
+  8658 : 'rArr',
+  8659 : 'dArr',
+  8660 : 'hArr',
+  8704 : 'forall',
+  8706 : 'part',
+  8707 : 'exist',
+  8709 : 'empty',
+  8711 : 'nabla',
+  8712 : 'isin',
+  8713 : 'notin',
+  8715 : 'ni',
+  8719 : 'prod',
+  8721 : 'sum',
+  8722 : 'minus',
+  8727 : 'lowast',
+  8730 : 'radic',
+  8733 : 'prop',
+  8734 : 'infin',
+  8736 : 'ang',
+  8743 : 'and',
+  8744 : 'or',
+  8745 : 'cap',
+  8746 : 'cup',
+  8747 : 'int',
+  8756 : 'there4',
+  8764 : 'sim',
+  8773 : 'cong',
+  8776 : 'asymp',
+  8800 : 'ne',
+  8801 : 'equiv',
+  8804 : 'le',
+  8805 : 'ge',
+  8834 : 'sub',
+  8835 : 'sup',
+  8836 : 'nsub',
+  8838 : 'sube',
+  8839 : 'supe',
+  8853 : 'oplus',
+  8855 : 'otimes',
+  8869 : 'perp',
+  8901 : 'sdot',
+  8968 : 'lceil',
+  8969 : 'rceil',
+  8970 : 'lfloor',
+  8971 : 'rfloor',
+  9001 : 'lang',
+  9002 : 'rang',
+  9674 : 'loz',
+  9824 : 'spades',
+  9827 : 'clubs',
+  9829 : 'hearts',
+  9830 : 'diams',
+  338 : 'OElig',
+  339 : 'oelig',
+  352 : 'Scaron',
+  353 : 'scaron',
+  376 : 'Yuml',
+  710 : 'circ',
+  732 : 'tilde',
+  8194 : 'ensp',
+  8195 : 'emsp',
+  8201 : 'thinsp',
+  8204 : 'zwnj',
+  8205 : 'zwj',
+  8206 : 'lrm',
+  8207 : 'rlm',
+  8211 : 'ndash',
+  8212 : 'mdash',
+  8216 : 'lsquo',
+  8217 : 'rsquo',
+  8218 : 'sbquo',
+  8220 : 'ldquo',
+  8221 : 'rdquo',
+  8222 : 'bdquo',
+  8224 : 'dagger',
+  8225 : 'Dagger',
+  8240 : 'permil',
+  8249 : 'lsaquo',
+  8250 : 'rsaquo',
+  8364 : 'euro'
+};
+
+var reverseEntities;
+// Match on named entities as well as numeric/hex entities as well,
+// covering range from &something; &Another; &#1234; &#x22; &#X2F;
+// http://www.whatwg.org/specs/web-apps/current-work/multipage/syntax.html#character-references
+var entityRegExp = /\&([#a-zA-Z0-9]+);/g;
+
+function makeReverseEntities () {
+  reverseEntities = {};
+  Object.keys(entities).forEach(function (key) {
+    reverseEntities[entities[key]] = key;
+  });
+}
+
+/**
+ * Escapes HTML characters like [<>"'&] in the text,
+ * leaving existing HTML entities intact.
+ */
+function escapeHTMLTextKeepingExistingEntities(text) {
+  return text.replace(/[<>"']|&(?![#a-zA-Z0-9]+;)/g, function(c) {
+    return '&#' + c.charCodeAt(0) + ';';
+  });
+}
+
+exports.unescapeHTMLEntities = function unescapeHTMLEntities(text) {
+  return text.replace(entityRegExp, function (match, ref) {
+    var converted = '';
+    if (ref.charAt(0) === '#') {
+      var secondChar = ref.charAt(1);
+      if (secondChar === 'x' || secondChar === 'X') {
+        // hex
+        converted = String.fromCharCode(parseInt(ref.substring(2), 16));
+      } else {
+        // base 10 reference
+        converted = String.fromCharCode(parseInt(ref.substring(1), 10));
+      }
+    } else {
+      // a named character reference
+      // build up reverse entities on first use.
+      if (!reverseEntities)
+        makeReverseEntities();
+
+      if (reverseEntities.hasOwnProperty(ref))
+        converted = String.fromCharCode(reverseEntities[ref]);
+    }
+    return converted;
+  });
+};
+
+/**
+ * Renders text content safe for injecting into HTML by
+ * replacing all characters which could be used to create HTML elements.
+ */
+exports.escapePlaintextIntoElementContext = function (text) {
+  return text.replace(/[&<>"'\/]/g, function(c) {
+    var code = c.charCodeAt(0);
+    return '&' + (entities[code] || '#' + code) + ';';
+  });
+}
+
+/**
+ * Escapes all characters with ASCII values less than 256, other than
+ * alphanumeric characters, with the &#xHH; format to prevent
+ * switching out of the attribute.
+ */
+exports.escapePlaintextIntoAttribute = function (text) {
+  return text.replace(/[\u0000-\u002F\u003A-\u0040\u005B-\u0060\u007B-\u0100]/g, function(c) {
+    var code = c.charCodeAt(0);
+    return '&' + (entities[code] || '#' + code) + ';';
+  });
+}
+
+
+}); // end define
+;
+/**
+ * Process text/html for message body purposes.  Specifically:
+ *
+ * - sanitize HTML (using bleach.js): discard illegal markup entirely, render
+ *   legal but 'regulated' markup inert (ex: links to external content).
+ * - TODO: perform normalization of quote markup from different clients into
+ *   blockquotes, like how Thunderbird conversations does it.
+ * - snippet generation: Try and generate a usable snippet string from something
+ *   that is not a quote.
+ *
+ * We may eventually try and perform more detailed analysis like `quotechew.js`
+ * does with structured markup, potentially by calling out to quotechew, but
+ * that's a tall order to get right, so it's mightily postponed.
+ **/
+
+define('mailapi/htmlchew',
+  [
+    'exports',
+    'bleach'
+  ],
+  function(
+    exports,
+    $bleach
+  ) {
+
+/**
+ * Whitelisted HTML tags list. Currently from nsTreeSanitizer.cpp which credits
+ * Mark Pilgrim and Sam Ruby for its own initial whitelist.
+ *
+ * IMPORTANT THUNDERBIRD NOTE: Thunderbird only engages its sanitization logic
+ * when processing mailto URIs, when the non-default
+ * "view | message body as | simple html" setting is selected, or when
+ * displaying spam messages.  Accordingly, the settings are pretty strict
+ * and not particularly thought-out.  Non-CSS presentation is stripped, which
+ * is pretty much the lingua franca of e-mail.  (Thunderbird itself generates
+ * font tags, for example.)
+ *
+ * Some things are just not in the list at all:
+ * - SVG: Thunderbird nukes these itself because it forces
+ *   SanitizerCidEmbedsOnly which causes flattening of everything in the SVG
+ *   namespace.
+ *
+ * Tags that we are opting not to include will be commented with a reason tag:
+ * - annoying: This thing is ruled out currently because it only allows annoying
+ *   things to happen *given our current capabilities*.
+ * - scripty: This thing requires scripting to make anything happen, and we do
+ *   not allow scripting.
+ * - forms: We have no UI to expose the target of a form right now, so it's
+ *   not safe.  Thunderbird displays a scam warning, which isn't realy a big
+ *   help, but it's something.  Because forms are largely unsupported or just
+ *   broken in many places, they are rarely used, so we are turning them off
+ *   entirely.
+ * - non-body: previously killed as part of the parse process because we were
+ *   assigning to innerHTML rather than creating a document with the string in
+ *   it.  We could change this up in a future bug now.
+ * - dangerous: The semantics of the tag are intentionally at odds with our
+ *   goals and/or are extensible.  (ex: link tag.)  Our callbacks could be
+ *   used to only let through okay things.
+ * - interactive-ui: A cross between scripty and forms, things like (HTML5)
+ *   menu and command imply some type of mutation that requires scripting.
+ *   They also are frequently very attribute-heavy.
+ * - svg: it's SVG, we don't support it yet!
+ */
+var LEGAL_TAGS = [
+  'a', 'abbr', 'acronym', 'area', 'article', 'aside',
+  // annoying: 'audio',
+  'b',
+  'bdi', 'bdo', // (bidirectional markup stuff)
+  'big', 'blockquote',
+  // implicitly-nuked: 'body'
+  'br',
+  // forms: 'button',
+  // scripty: canvas
+  'caption',
+  'center',
+  'cite', 'code', 'col', 'colgroup',
+  // interactive-ui: 'command',
+  // forms: 'datalist',
+  'dd', 'del', 'details', 'dfn', 'dir', 'div', 'dl', 'dt',
+  'em',
+  // forms: 'fieldset' (but allowed by nsTreeSanitizer)
+  'figcaption', 'figure',
+  'font',
+  'footer',
+  // forms: 'form',
+  'h1', 'h2', 'h3', 'h4', 'h5', 'h6',
+  // non-body: 'head'
+  'header', 'hgroup', 'hr',
+  // non-body: 'html'
+  'i', 'img',
+  // forms: 'input',
+  'ins', // ("represents a range of text that has been inserted to a document")
+  'kbd', // ("The kbd element represents user input")
+  'label', 'legend', 'li',
+  // dangerous: link (for CSS styles
+  /* link supports many types, none of which we want, some of which are
+   * risky: http://dev.w3.org/html5/spec/links.html#linkTypes. Specifics:
+   * - "stylesheet": This would be okay for cid links, but there's no clear
+   *   advantage over inline styles, so we forbid it, especially as supporting
+   *   it might encourage other implementations to dangerously support link.
+   * - "prefetch": Its whole point is de facto information leakage.
+   */
+  'listing', // (deprecated, like "pre")
+  'map', 'mark',
+  // interactive-ui: 'menu', 'meta', 'meter',
+  'nav',
+  'nobr', // (deprecated "white-space:nowrap" equivalent)
+  'noscript',
+  'ol',
+  // forms: 'optgroup',
+  // forms: 'option',
+  'output', // (HTML5 draft: "result of a calculation in a form")
+  'p', 'pre',
+  // interactive-ui: 'progress',
+  'q',
+  /* http://www.w3.org/TR/ruby/ is a pronounciation markup that is not directly
+   * supported by gecko at this time (although there is a Firefox extension).
+   * All of 'rp', 'rt', and 'ruby' are ruby tags.  The spec also defines 'rb'
+   * and 'rbc' tags that nsTreeSanitizer does not whitelist, however.
+   */
+  'rp', 'rt', 'ruby',
+  's', 'samp', 'section',
+  // forms: 'select',
+  'small',
+  // annoying?: 'source',
+  'span', 'strike', 'strong',
+  'style',
+  'sub', 'summary', 'sup',
+  // svg: 'svg', NB: this lives in its own namespace
+  'table', 'tbody', 'td',
+  // forms: 'textarea',
+  'tfoot', 'th', 'thead', 'time',
+  'title', // XXX does this mean anything outside head?
+  'tr',
+  // annoying?: 'track'
+  'tt',
+  'u', 'ul', 'var',
+  // annoying: 'video',
+  'wbr' // (HTML5 draft: line break opportunity)
+];
+
+/**
+ * Tags whose children should be removed along with the tag itself, rather than
+ * splicing the children into the position originally occupied by the parent.
+ *
+ * We do this for:
+ * - forms; see `LEGAL_TAGS` for the rationale.  Note that we don't bother
+ *   including children that should already be nuked by PRUNE_TAGS.  For
+ *   example, 'option' and 'optgroup' only make sense under 'select' or
+ *   'datalist', so we need not include them.  This means that if the tags
+ *   are used in nonsensical positions, they will have their contents
+ *   merged into the document text, but that's not a major concern.
+ * - non-body: don't have stuff from the header show up like it's part of the
+ *   body!  For now we do want <style> tags to fall out, but we want <title>
+ *   to not show up, etc.
+ * - 'script': no one wants to read the ignored JS code!
+ * Note that bleach.js now is aware of the special nature of 'script' and
+ * 'style' tags, so putting them in prune is not strictly required.
+ */
+var PRUNE_TAGS = [
+  'button', // (forms)
+  'datalist', // (forms)
+  'script', // (script)
+  'select', // (forms)
+  'svg', // (svg)
+  'title', // (non-body)
+];
+
+/**
+ * What attributes to allow globally and on specific tags.
+ *
+ * Forbidden marker names:
+ * - URL-like: The attribute can contain URL's and we don't care enough to
+ *   sanitize the contents right now.
+ * - sanitized: We manually do something with the attribute in our processing
+ *   logic.
+ * - specific: The attribute is explicitly named on the relevant element types.
+ * - unsupported: Gecko ignores the attribute and there is no chance of
+ *   standardization, so just strip it.
+ * - microformat: we can't do anything with microformats right now, save some
+ *   space.
+ * - awkward: It's not dangerous, but it's not clear how it could have useful
+ *   semantics.
+ */
+var LEGAL_ATTR_MAP = {
+  '*': [
+    'abbr', // (tables: removed from HTML5)
+    // forms: 'accept', 'accept-charset',
+    // interactive-ui: 'accesskey',
+    // forms: 'action',
+    'align', // (pres)
+    'alt', // (fallback content)
+    // forms: 'autocomplete', 'autofocus',
+    // annoying: 'autoplay',
+    'axis', // (tables: removed from HTML5)
+    // URL-like: 'background',
+    'bgcolor', 'border', // (pres)
+    'cellpadding', 'cellspacing', // (pres)
+    // unsupported: 'char',
+    'charoff', // (tables)
+    // specific: 'charset'
+    // forms, interactive-ui: 'checked',
+    // URL-like: 'cite'
+    'class', 'clear', 'color', // (pres)
+    'cols', 'colspan', // (tables)
+    'compact', // (pres)
+    // dangerous: 'content', (meta content refresh is bad.)
+    // interactive-ui: 'contenteditable', (we already use this ourselves!)
+    // interactive-ui: 'contextmenu',
+    // annoying: 'controls', (media)
+    'coords', // (area image map)
+    'datetime', // (ins, del, time semantic markups)
+    // forms: 'disabled',
+    'dir', // (rtl)
+    // interactive-ui: 'draggable',
+    // forms: 'enctype',
+    'face', // (pres)
+    // forms: 'for',
+    'frame', // (tables)
+    'headers', // (tables)
+    'height', // (layout)
+    // interactive-ui: 'hidden', 'high',
+    // sanitized: 'href',
+    // specific: 'hreflang',
+    'hspace', // (pres)
+    // dangerous: 'http-equiv' (meta refresh, maybe other trickiness)
+    // interactive-ui: 'icon',
+    'id', // (pres; white-listed for style targets)
+    // specific: 'ismap', (area image map)
+    // microformat: 'itemid', 'itemprop', 'itemref', 'itemscope', 'itemtype',
+    // annoying: 'kind', (media)
+    // annoying, forms, interactive-ui: 'label',
+    'lang', // (language support)
+    // forms: 'list',
+    // dangerous: 'longdesc', (link to a long description, html5 removed)
+    // annoying: 'loop',
+    // interactive-ui: 'low',
+    // forms, interactive-ui: 'max',
+    // forms: 'maxlength',
+    'media', // (media-query for linky things; safe if links are safe)
+    // forms: 'method',
+    // forms, interactive-ui: 'min',
+    // unsupported: 'moz-do-not-send', (thunderbird internal composition)
+    // forms: 'multiple',
+    // annoying: 'muted',
+    // forms, interactive-ui: 'name', (although pretty safe)
+    'nohref', // (image maps)
+    // forms: 'novalidate',
+    'noshade', // (pres)
+    'nowrap', // (tables)
+    'open', // (for "details" element)
+    // interactive-ui: 'optimum',
+    // forms: 'pattern', 'placeholder',
+    // annoying: 'playbackrate',
+    'pointsize', // (pres)
+    // annoying:  'poster', 'preload',
+    // forms: 'prompt',
+    'pubdate', // ("time" element)
+    // forms: 'radiogroup', 'readonly',
+    // dangerous: 'rel', (link rel, a rel, area rel)
+    // forms: 'required',
+    // awkward: 'rev' (reverse link; you can't really link to emails)
+    'reversed', // (pres? "ol" reverse numbering)
+    // interactive-ui: 'role', We don't want a screen reader making the user
+    //   think that part of the e-mail is part of the UI.  (WAI-ARIA defines
+    //   "accessible rich internet applications", not content markup.)
+    'rows', 'rowspan', 'rules', // (tables)
+    // sanitized: 'src',
+    'size', // (pres)
+    'scope', // (tables)
+    'scoped', // (pres; on "style" elem)
+    // forms: 'selected',
+    'shape', // (image maps)
+    'span', // (tables)
+    // interactive-ui: 'spellcheck',
+    // sanitized, dangerous: 'src'
+    // annoying: 'srclang',
+    'start', // (pres? "ol" numbering)
+    'summary', // (tables accessibility)
+    'style', // (pres)
+    // interactive-ui: 'tabindex',
+    // dangerous: 'target', (specifies a browsing context, but our semantics
+    //   are extremely clear and don't need help.)
+    'title', // (advisory)
+    // specific, dangerous: type (various, but mime-type for links is not the
+    //   type of thing we would ever want to propagate or potentially deceive
+    //   the user with.)
+    'valign', // (pres)
+    'value', // (pres? "li" override for "ol"; various form uses)
+    'vspace', // (pres)
+    'width', // (layout)
+    // forms: 'wrap',
+  ],
+  'a': ['ext-href', 'hreflang'],
+  'area': ['ext-href', 'hreflang'],
+  // these are used by our quoting and Thunderbird's quoting
+  'blockquote': ['cite', 'type'],
+  'img': ['cid-src', 'ext-src', 'ismap', 'usemap'],
+  // This may only end up being used as a debugging thing, but let's let charset
+  // through for now.
+  'meta': ['charset'],
+  'ol': ['type'], // (pres)
+  'style': ['type'],
+};
+
+/**
+ * CSS Style rules to support.
+ *
+ * nsTreeSanitizer is super lazy about style binding and does not help us out.
+ * What it does is nuke all rule types except NAMESPACE (@namespace), FONT_FACE
+ * (@font-face), and STYLE rules (actual styling).  This means nuking CHARSET
+ * (@charset to specify the encoding of the stylesheet if the server did not
+ * provide it), IMPORT (@import to reference other stylesheet files), MEDIA
+ * (@media media queries), PAGE (@page page box info for paged media),
+ * MOZ_KEYFRAMES, MOZ_KEYFRAME, SUPPORTS (@supports provides support for rules
+ * conditioned on browser support, but is at risk.)  The only style directive it
+ * nukes is "-moz-binding" which is the XBL magic and considered dangerous.
+ *
+ * Risks: Anything that takes a url() is dangerous insofar as we need to
+ * sanitize the url.  XXX for now we just avoid any style that could potentially
+ * hold a URI.
+ *
+ * Good news: We always cram things into an iframe, so we don't need to worry
+ * about clever styling escaping out into our UI.
+ *
+ * New reasons not to allow:
+ * - animation: We don't want or need animated wackiness.
+ * - slow: Doing the thing is slow!
+ */
+var LEGAL_STYLES = [
+  // animation: animation*
+  // URI-like: background, background-image
+  'background-color',
+  // NB: border-image is not set by the 'border' aliases
+  'border',
+  'border-bottom', 'border-bottom-color', 'border-bottom-left-radius',
+  'border-bottom-right-radius', 'border-bottom-style', 'border-bottom-width',
+  'border-color',
+  // URI-like: border-image*
+  'border-left', 'border-left-color', 'border-left-style', 'border-left-width',
+  'border-radius',
+  'border-right', 'border-right-color', 'border-right-style',
+  'border-right-width',
+  'border-style',
+  'border-top', 'border-top-color', 'border-top-left-radius',
+  'border-top-right-radius', 'border-top-style', 'border-top-width',
+  'border-width',
+  // slow: box-shadow
+  'clear',
+  'color',
+  'display',
+  'float',
+  'font-family',
+  'font-size',
+  'font-style',
+  'font-weight',
+  'height',
+  'line-height',
+  // URI-like: list-style, list-style-image
+  'list-style-position',
+  'list-style-type',
+  'margin', 'margin-bottom', 'margin-left', 'margin-right', 'margin-top',
+  'padding', 'padding-bottom', 'padding-left', 'padding-right', 'padding-top',
+  'text-align', 'text-align-last',
+  'text-decoration', 'text-decoration-color', 'text-decoration-line',
+  'text-decoration-style', 'text-indent',
+  'vertical-align',
+  'white-space',
+  'width',
+  'word-break', 'word-spacing', 'word-wrap',
+];
+
+/**
+ * The regular expression to detect nodes that should be passed to stashLinks.
+ *
+ * ignore-case is not required; the value is checked against the lower-cased tag.
+ */
+var RE_NODE_NEEDS_TRANSFORM = /^(?:a|area|img)$/;
+
+var RE_CID_URL = /^cid:/i;
+var RE_HTTP_URL = /^http(?:s)?/i;
+var RE_MAILTO_URL = /^mailto:/i;
+
+var RE_IMG_TAG = /^img$/;
+
+function getAttributeFromList(attrs, name) {
+  var len = attrs.length;
+  for (var i = 0; i < len; i++) {
+    var attr = attrs[i];
+    if (attr.name.toLowerCase() === name) {
+      return attr;
+    }
+  }
+  return null;
+}
+
+/**
+ * Transforms src tags, ensure that links are http and transform them too so
+ * that they don't actually navigate when clicked on but we can hook them.  (The
+ * HTML display iframe is not intended to navigate; we just want to trigger the
+ * browser.
+ */
+function stashLinks(lowerTag, attrs) {
+  var classAttr;
+  // - img: src
+  if (RE_IMG_TAG.test(lowerTag)) {
+    // filter out things we might write to, also find the 'class attr'
+    attrs = attrs.filter(function(attr) {
+      switch (attr.name.toLowerCase()) {
+        case 'cid-src':
+        case 'ext-src':
+          return false;
+        case 'class':
+          classAttr = attr;
+        default:
+          return true;
+      }
+    });
+
+    var srcAttr = getAttributeFromList(attrs, 'src');
+    if (srcAttr) {
+      if (RE_CID_URL.test(srcAttr.escaped)) {
+        srcAttr.name = 'cid-src';
+        if (classAttr)
+          classAttr.escaped += ' moz-embedded-image';
+        else
+          attrs.push({ name: 'class', escaped: 'moz-embedded-image' });
+        // strip the cid: bit, it is necessarily there and therefore redundant.
+        srcAttr.escaped = srcAttr.escaped.substring(4);
+      }
+      else if (RE_HTTP_URL.test(srcAttr.escaped)) {
+        srcAttr.name = 'ext-src';
+        if (classAttr)
+          classAttr.escaped += ' moz-external-image';
+        else
+          attrs.push({ name: 'class', escaped: 'moz-external-image' });
+      }
+    }
+  }
+  // - a, area: href
+  else {
+    // filter out things we might write to, also find the 'class attr'
+    attrs = attrs.filter(function(attr) {
+      switch (attr.name.toLowerCase()) {
+        case 'cid-src':
+        case 'ext-src':
+          return false;
+        case 'class':
+          classAttr = attr;
+        default:
+          return true;
+      }
+    });
+    var linkAttr = getAttributeFromList(attrs, 'href');
+    if (linkAttr) {
+      var link = linkAttr.escaped;
+      if (RE_HTTP_URL.test(link) ||
+          RE_MAILTO_URL.test(link)) {
+
+        linkAttr.name = 'ext-href';
+        if (classAttr)
+          classAttr.escaped += ' moz-external-link';
+        else
+          attrs.push({ name: 'class', escaped: 'moz-external-link' });
+      }
+      else {
+        // paranoia; no known benefit if this got through
+        attrs.splice(attrs.indexOf(linkAttr), 1);
+      }
+    }
+  }
+  return attrs;
+}
+
+var BLEACH_SETTINGS = {
+  tags: LEGAL_TAGS,
+  strip: true,
+  stripComments: true,
+  prune: PRUNE_TAGS,
+  attributes: LEGAL_ATTR_MAP,
+  styles: LEGAL_STYLES,
+  asNode: true,
+  callbackRegexp: RE_NODE_NEEDS_TRANSFORM,
+  callback: stashLinks
+};
+
+var BLEACH_SNIPPET_SETTINGS = {
+  tags: [],
+  strip: true,
+  stripComments: true,
+  prune: [
+    'style',
+    'button', // (forms)
+    'datalist', // (forms)
+    'script', // (script)
+    'select', // (forms)
+    'svg', // (svg)
+    'title' // (non-body)
+  ],
+  asNode: true,
+  maxLength: 100
+};
+
+/**
+ * @args[
+ *   @param[htmlString String]{
+ *     An unsanitized HTML string.  The HTML content can be a fully valid HTML
+ *     document with 'html' and 'body' tags and such, but most of that extra
+ *     structure will currently be discarded.
+ *
+ *     In the future we may try and process the body and such correctly, but for
+ *     now we don't.  This is consistent with many webmail clients who ignore
+ *     style tags in the head, etc.
+ *   }
+ * ]
+ * @return[HtmlString]{
+ *   The sanitized HTML string wrapped into a div container.
+ * }
+ */
+exports.sanitizeAndNormalizeHtml = function sanitizeAndNormalize(htmlString) {
+  return $bleach.clean(htmlString, BLEACH_SETTINGS);
+};
+
+/**
+ * Derive snippet text from the an HTML string. It will also sanitize it.
+ * Note that it unescapes HTML enttities, so best to only use this output
+ * in textContent cases.
+ */
+exports.generateSnippet = function generateSnippet(htmlString) {
+  return $bleach.unescapeHTMLEntities($bleach.clean(htmlString,
+                                                    BLEACH_SNIPPET_SETTINGS));
+};
+
+var BLEACH_SEARCHABLE_TEXT_WITH_QUOTES_SETTINGS = {
+  tags: [],
+  strip: true,
+  stripComments: true,
+  prune: [
+    'style',
+    'button', // (forms)
+    'datalist', // (forms)
+    'script', // (script)
+    'select', // (forms)
+    'svg', // (svg)
+    'title' // (non-body)
+  ],
+  asNode: true,
+};
+
+var BLEACH_SEARCHABLE_TEXT_WITHOUT_QUOTES_SETTINGS = {
+  tags: [],
+  strip: true,
+  stripComments: true,
+  prune: [
+    'style',
+    'button', // (forms)
+    'datalist', // (forms)
+    'script', // (script)
+    'select', // (forms)
+    'svg', // (svg)
+    'title', // (non-body),
+    // specific to getting rid of quotes:
+    'blockquote'
+  ],
+  asNode: true,
+};
+
+
+/**
+ * Produce a textual version of the body of the e-mail suitable for search
+ * purposes.  This is basically the same thing as generateSnippet but without a
+ * length limit applied and with the ability to either include quoted text or
+ * not include quoted text.  We do process the entire document in a go and
+ * return the entire results, so this could be fairly inefficient from a
+ * memory/time perspective.
+ *
+ * The following potential enhancements could be fairly good ideas:
+ * - Avoid processing the entire HTML document by passing the search string in
+ *   or using a generator-type implementation to yield incremental string hunks.
+ * - Generate a semantic representation similar/identical to the one used by
+ *   quotechew (at least for this searchable text mode.)
+ */
+exports.generateSearchableTextVersion = function(htmlString, includeQuotes) {
+  var settings;
+  if (includeQuotes) {
+    settings = BLEACH_SEARCHABLE_TEXT_WITH_QUOTES_SETTINGS;
+  }
+  else {
+    settings = BLEACH_SEARCHABLE_TEXT_WITHOUT_QUOTES_SETTINGS;
+  }
+  var cleaned = $bleach.clean(htmlString, settings);
+  return $bleach.unescapeHTMLEntities(cleaned);
+};
+
+/**
+ * Wrap text/plain content into a serialized HTML string safe for insertion
+ * via innerHTML.
+ *
+ * By default we wrap everything in a 'div' tag with 'br' indicating newlines.
+ * Alternately, we could use 'white-space: pre-wrap' if we were more confident
+ * about recipients having sufficient CSS support and our own desire to have
+ * things resemble text/plain.
+ *
+ */
+exports.wrapTextIntoSafeHTMLString = function(text, wrapTag,
+                                              transformNewlines, attrs) {
+  if (transformNewlines === undefined) {
+    transformNewlines = true;
+  }
+
+  wrapTag = wrapTag || 'div';
+
+  text = $bleach.escapePlaintextIntoElementContext(text);
+  text = transformNewlines ? text.replace(/\n/g, '<br/>') : text;
+
+  var attributes = '';
+  if (attrs) {
+    var len = attrs.length;
+    for (var i = 0; i < len; i += 2) {
+      attributes += ' ' + attrs[i] + '="' +
+        $bleach.escapePlaintextIntoAttribute(attrs[i + 1]) + '"';
+    }
+  }
+
+  return '<' + wrapTag + attributes + '>' + text + '</' + wrapTag + '>';
+};
+
+var RE_QUOTE_CHAR = /"/g;
+
+/**
+ * Make an HTML attribute value safe.
+ */
+exports.escapeAttrValue = function(s) {
+  return s.replace(RE_QUOTE_CHAR, '&quot;');
+};
+
+}); // end define
+;
 /**
  * Searchfilters provide for local searching by checking each message against
  * one or more tests.  This is similar to Thunderbird's non-global search
@@ -9806,26 +12770,46 @@ return {
  * - Recipients
  * - Subject
  * - Body, allows ignoring quoted bits
- *
- * XXX currently all string searching uses indexOf; we at the very least should
- * build a regexp that is configured to ignore case.
  **/
 
 define('mailapi/searchfilter',
   [
     'rdcommon/log',
+    './util',
     './syncbase',
     './date',
+    './htmlchew',
     'module',
     'exports'
   ],
   function(
     $log,
+    $util,
     $syncbase,
     $date,
+    htmlchew,
     $module,
     exports
   ) {
+var BEFORE = $date.BEFORE,
+    ON_OR_BEFORE = $date.ON_OR_BEFORE,
+    SINCE = $date.SINCE,
+    STRICTLY_AFTER = $date.STRICTLY_AFTER;
+var bsearchMaybeExists = $util.bsearchMaybeExists,
+    bsearchForInsert = $util.bsearchForInsert;
+
+/**
+ * cmpHeaderYoungToOld with matched-header unwrapping
+ */
+function cmpMatchHeadersYoungToOld(aMatch, bMatch) {
+  var a = aMatch.header, b = bMatch.header;
+  var delta = b.date - a.date;
+  if (delta)
+    return delta;
+  // favor larger UIDs because they are newer-ish.
+  return b.id - a.id;
+
+}
 
 /**
  * This internal function checks if a string or a regexp matches an input
@@ -9984,13 +12968,26 @@ function snippetMatchHelper(str, start, length, contextBefore, contextAfter,
   if (contextBefore > start)
     contextBefore = start;
   var offset = str.indexOf(' ', start - contextBefore);
-  if (offset === -1)
-    offset = 0;
-  if (offset >= start)
+  // Just fragment the preceding word if there was no match whatsoever or the
+  // whitespace match happened preceding our word or anywhere after it.
+  if (offset === -1 || offset >= (start - 1)) {
     offset = start - contextBefore;
-  var endIdx = str.lastIndexOf(' ', start + length + contextAfter);
-  if (endIdx <= start + length)
-    endIdx = start + length + contextAfter;
+  }
+  else {
+    // do not start on the space character
+    offset++;
+  }
+
+  var endIdx;
+  if (start + length + contextAfter >= str.length) {
+    endIdx = str.length;
+  }
+  else {
+    endIdx = str.lastIndexOf(' ', start + length + contextAfter - 1);
+    if (endIdx <= start + length) {
+      endIdx = start + length + contextAfter;
+    }
+  }
   var snippet = str.substring(offset, endIdx);
 
   return {
@@ -10083,7 +13080,7 @@ BodyFilter.prototype = {
         contextBefore = this.contextBefore, contextAfter = this.contextAfter,
         matches = [],
         matchQuotes = this.matchQuotes,
-        idx;
+        idx, ret;
 
     for (var iBodyRep = 0; iBodyRep < body.bodyReps.length; iBodyRep++) {
       var bodyType = body.bodyReps[iBodyRep].type,
@@ -10101,82 +13098,36 @@ BodyFilter.prototype = {
             continue;
 
           for (idx = 0; idx < block.length && matches.length < stopAfter;) {
-            var ret = matchRegexpOrString(phrase, block, idx);
-            if (!ret)
+            ret = matchRegexpOrString(phrase, block, idx);
+            if (!ret) {
               break;
-            if (repPath === null)
+            }
+            if (repPath === null) {
               repPath = [iBodyRep, iRep];
-            matches.push(snippetMatchHelper(block, ret.index, ret[0].length,
-                                            contextBefore, contextAfter,
-                                            repPath));
+            }
+            matches.push(snippetMatchHelper(
+              block, idx + ret.index, ret[0].length,
+              contextBefore, contextAfter,
+              repPath));
             idx += ret.index + ret[0].length;
           }
         }
       }
       else if (bodyType === 'html') {
-        // NB: this code is derived from htmlchew.js' generateSnippet
-        // functionality.
-
-        // - convert the HMTL into a DOM tree
-        // We don't want our indexOf to run afoul of presentation logic.
-        var htmlPath = [iBodyRep, 0];
-        var htmlDoc = document.implementation.createHTMLDocument(''),
-            rootNode = htmlDoc.createElement('div');
-        rootNode.innerHTML = bodyRep;
-
-        var node = rootNode.firstChild, done = false;
-        while (!done) {
-          if (node.nodeType === ELEMENT_NODE) {
-            switch (node.tagName.toLowerCase()) {
-              // - Things that can't contain useful text.
-              // The style does not belong in the snippet!
-              case 'style':
-                break;
-
-              case 'blockquote':
-                // fall-through if matchQuotes
-                if (!matchQuotes)
-                  break;
-              default:
-                if (node.firstChild) {
-                  node = node.firstChild;
-                  htmlPath.push(0);
-                  continue;
-                }
-                break;
-            }
+        var searchableText = htmlchew.generateSearchableTextVersion(
+          bodyRep, this.matchQuotes);
+        for (idx = 0; idx < bodyRep.length && matches.length < stopAfter;) {
+          ret = matchRegexpOrString(phrase, searchableText, idx);
+          if (!ret) {
+            break;
           }
-          else if (node.nodeType === TEXT_NODE) {
-            // XXX the snippet generator normalizes whitespace here to avoid
-            // being overwhelmed by ridiculous whitespace.  This is not quite
-            // as much a problem for us, but it would be useful if the
-            // sanitizer layer normalized whitespace so neither of us has to
-            // worry about it.
-            var nodeText = node.data;
-
-            var ret = matchRegexpOrString(phrase, nodeText, 0);
-            if (ret) {
-              matches.push(
-                snippetMatchHelper(nodeText, ret.index, ret[0].length,
-                                   contextBefore, contextAfter,
-                                   htmlPath.concat()));
-              if (matches.length >= stopAfter)
-                break;
-            }
-          }
-
-          while (!node.nextSibling) {
-            node = node.parentNode;
-            htmlPath.pop();
-            if (node === rootNode) {
-              done = true;
-              break;
-            }
-          }
-          if (!done) {
-            node = node.nextSibling;
-            htmlPath[htmlPath.length - 1]++;
-          }
+          // note: because we heavily discard DOM structure, we are unable to
+          // generate a useful path.  The good news is we don't use the path
+          // anywhere at this time, so it's not particularly a big deal.
+          matches.push(snippetMatchHelper(
+            searchableText, idx + ret.index, ret[0].length,
+            contextBefore, contextAfter, null));
+          idx += ret.index + ret[0].length;
         }
       }
     }
@@ -10200,6 +13151,12 @@ function MessageFilterer(filters) {
   this.filters = filters;
   this.bodiesNeeded = false;
 
+  /**
+   * How many headers have we tried to match against?  This is for unit tests.
+   */
+  this.messagesChecked = 0;
+
+
   for (var i = 0; i < filters.length; i++) {
     var filter = filters[i];
     if (filter.needsBody)
@@ -10214,6 +13171,8 @@ MessageFilterer.prototype = {
    * are defined by the filterers in use.
    */
   testMessage: function(header, body) {
+    this.messagesChecked++;
+
     //console.log('sf: testMessage(', header.suid, header.author.address,
     //            header.subject, 'body?', !!body, ')');
     var matched = false, matchObj = {};
@@ -10255,6 +13214,15 @@ console.log('sf: creating SearchSlice:', phrase);
   // These correspond to the range of headers that we have searched to generate
   // the current set of matched headers.  Our matches will always be fully
   // contained by this range.
+  //
+  // This range can and will shrink when reqNoteRanges is called.  Currently we
+  // shrink to the first/last remaining matches.  Strictly speaking, this is too
+  // aggressive.  The optimal shrink constraint would be to pick the message
+  // adjacent to the first matches we are discarding so that growing by one
+  // message would immediately re-find the message.  However it would be even
+  // MORE efficient to just maintain a compact list of messages that have
+  // matched that we never forget, so we'll just do that when we're feeling all
+  // fancy in the future.
   this.startTS = null;
   this.startUID = null;
   this.endTS = null;
@@ -10284,26 +13252,86 @@ console.log('sf: creating SearchSlice:', phrase);
   this._bound_gotOlderMessages = this._gotMessages.bind(this, 1);
   this._bound_gotNewerMessages = this._gotMessages.bind(this, -1);
 
-  this.headers = [];
   this.desiredHeaders = $syncbase.INITIAL_FILL_SIZE;
-  // Fetch as many headers as we want in our results; we probably will have
-  // less than a 100% hit-rate, but there isn't much savings from getting the
-  // extra headers now, so punt on those.
-  this._storage.getMessagesInImapDateRange(
-    0, null, this.desiredHeaders, this.desiredHeaders,
-    this._gotMessages.bind(this, 1));
+  this.reset();
 }
 exports.SearchSlice = SearchSlice;
 SearchSlice.prototype = {
+  /**
+   * We are a filtering search slice.  To reduce confusion, we still call this
+   * search.
+   */
+  type: 'search',
+
   set atTop(val) {
     this._bridgeHandle.atTop = val;
+  },
+  get atBottom() {
+    return this._bridgeHandle.atBottom;
   },
   set atBottom(val) {
     this._bridgeHandle.atBottom = val;
   },
+  set headerCount(val) {
+    if (this._bridgeHandle)
+      this._bridgeHandle.headerCount = val;
+    return val;
+  },
+
+  /**
+   * How many messages should we pretend exist when we haven't yet searched all
+   * of the folder?
+   *
+   * As a lazy search, we have no idea how many messages actually match a user's
+   * search.  We now assume a virtual scroll list that sizes itself based on
+   * knowing how many headers there are using headerCount and thus no longer
+   * really cares about atBottom (at least until we start automatically
+   * synchronizing new messages.)
+   *
+   * 1 is a pretty good value for this since it only takes 1 lied-about message
+   * to trigger us.  Also, the UI will show the "I'm still loading stuff!"
+   * fake message until we find something...
+   *
+   * TODO: Either stop lying or come up with a better rationale for this.  All
+   * we really want is for the UI to remember to ask us for more stuff, and all
+   * the UI probably wants is to show some type of search-specific string that
+   * says "Hey, I'm searching here!  Give it a minute!".  Not doing this right
+   * now because that results in all kinds of scope creep and such.
+   */
+  IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM: 1,
+
+  reset: function() {
+    // misnomer but simplifies cutting/pasting/etc.  Really an array of
+    // { header: header, matches: matchObj }
+    this.headers = [];
+    this.headerCount = 0;
+    // Track when we are still performing the initial database scan so that we
+    // can ignore dynamic additions/modifications.  The initial database scan
+    // is currently not clever enough to deal with concurrent manipulation, so
+    // we just ignore all such events.  This has an extremely low probability
+    // of resulting in false negatives.
+    this._loading = true;
+    this.startTS = null;
+    this.startUID = null;
+    this.endTS = null;
+    this.endUID = null;
+    // Fetch as many headers as we want in our results; we probably will have
+    // less than a 100% hit-rate, but there isn't much savings from getting the
+    // extra headers now, so punt on those.
+    this._storage.getMessagesInImapDateRange(
+      0, null, this.desiredHeaders, this.desiredHeaders,
+      this._gotMessages.bind(this, 1));
+  },
 
   _gotMessages: function(dir, headers, moreMessagesComing) {
-console.log('sf: gotMessages', headers.length);
+    if (!this._bridgeHandle) {
+      return;
+    }
+    // conditionally indent messages that are non-notable callbacks since we
+    // have more messages coming.  sanity measure for asuth for now.
+    var logPrefix = moreMessagesComing ? 'sf: ' : 'sf:';
+    console.log(logPrefix, 'gotMessages', headers.length, 'more coming?',
+                moreMessagesComing);
     // update the range of what we have seen and searched
     if (headers.length) {
       if (dir === -1) { // (more recent)
@@ -10322,11 +13350,16 @@ console.log('sf: gotMessages', headers.length);
     }
 
     var checkHandle = function checkHandle(headers, bodies) {
+      if (!this._bridgeHandle) {
+        return;
+      }
+
       // run a filter on these
       var matchPairs = [];
       for (i = 0; i < headers.length; i++) {
         var header = headers[i],
             body = bodies ? bodies[i] : null;
+        this._headersChecked++;
         var matchObj = this.filterer.testMessage(header, body);
         if (matchObj)
           matchPairs.push({ header: header, matches: matchObj });
@@ -10337,30 +13370,61 @@ console.log('sf: gotMessages', headers.length);
       var atBottom = this.atBottom = this._storage.headerIsOldestKnown(
                        this.startTS, this.startUID);
       var canGetMore = (dir === -1) ? !atTop : !atBottom;
+      var willHave = this.headers.length + matchPairs.length,
+          wantMore = !moreMessagesComing &&
+                     (willHave < this.desiredHeaders) &&
+                     canGetMore;
       if (matchPairs.length) {
-        var willHave = this.headers.length + matchPairs.length,
-            wantMore = !moreMessagesComing &&
-                       (willHave < this.desiredHeaders) &&
-                       canGetMore;
-console.log('sf: willHave', willHave, 'of', this.desiredHeaders, 'want more?', wantMore);
+        console.log(logPrefix, 'willHave', willHave, 'of', this.desiredHeaders,
+                    'want more?', wantMore);
         var insertAt = dir === -1 ? 0 : this.headers.length;
+        this._LOG.headersAppended(insertAt, matchPairs);
+
+        this.headers.splice.apply(this.headers,
+                                  [insertAt, 0].concat(matchPairs));
+        this.headerCount = this.headers.length +
+          (atBottom ? 0 : this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM);
+
         this._bridgeHandle.sendSplice(
           insertAt, 0, matchPairs, true,
           moreMessagesComing || wantMore);
-        this.headers.splice.apply(this.headers,
-                                  [insertAt, 0].concat(matchPairs));
-        if (wantMore)
-          this.reqGrow(dir, false);
+
+        if (wantMore) {
+          console.log(logPrefix, 'requesting more because want more');
+          this.reqGrow(dir, false, true);
+        }
+        else if (!moreMessagesComing) {
+          console.log(logPrefix, 'stopping (already reported), no want more.',
+                      'can get more?', canGetMore);
+          this._loading = false;
+          this.desiredHeaders = this.headers.length;
+        }
       }
+      // XXX this branch is largely the same as in the prior case except for
+      // specialization because the sendSplice call obviates the need to call
+      // sendStatus.  Consider consolidation.
       else if (!moreMessagesComing) {
+        // Update our headerCount, potentially reducing our headerCount by 1!
+        this.headerCount = this.headers.length +
+          (atBottom ? 0 : this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM);
+
         // If there aren't more messages coming, we either need to get more
         // messages (if there are any left in the folder that we haven't seen)
         // or signal completion.  We can use our growth function directly since
         // there are no state invariants that will get confused.
-        if (canGetMore)
-          this.reqGrow(dir, false);
-        else
+        if (wantMore) {
+          console.log(logPrefix,
+                      'requesting more because no matches but want more');
+          this.reqGrow(dir, false, true);
+        }
+        else {
+          console.log(logPrefix, 'stopping, no matches, no want more.',
+                      'can get more?', canGetMore);
           this._bridgeHandle.sendStatus('synced', true, false);
+          // We can now process dynamic additions/modifications
+          this._loading = false;
+          this.desiredHeaders = this.headers.length;
+        }
       }
       // (otherwise we need to wait for the additional messages to show before
       //  doing anything conclusive)
@@ -10368,20 +13432,31 @@ console.log('sf: willHave', willHave, 'of', this.desiredHeaders, 'want more?', w
 
     if (this.filterer.bodiesNeeded) {
       // To batch our updates to the UI, just get all the bodies then advance
-      // to the next stage of processing.  It would be nice
+      // to the next stage of processing.
       var bodies = [];
       var gotBody = function(body) {
-        if (!body)
-          console.log('failed to get a body for: ',
+        if (!body) {
+          console.log(logPrefix, 'failed to get a body for: ',
                       headers[bodies.length].suid,
                       headers[bodies.length].subject);
+        }
         bodies.push(body);
-        if (bodies.length === headers.length)
+        if (bodies.length === headers.length) {
           checkHandle(headers, bodies);
+        }
       };
       for (var i = 0; i < headers.length; i++) {
         var header = headers[i];
         this._storage.getMessageBody(header.suid, header.date, gotBody);
+      }
+      if (!headers.length) {
+        // To maintain consistent ordering for correctness we need to make sure
+        // we won't call checkHeaders (to trigger additional fetches if
+        // required) before any outstanding getMessageBody calls return.
+        // runAfterDeferredCalls guarantees us consistency with how
+        // getMessageBody operates in this scenario.
+        this._storage.runAfterDeferredCalls(
+          checkHandle.bind(null, headers, null));
       }
     }
     else {
@@ -10390,7 +13465,179 @@ console.log('sf: willHave', willHave, 'of', this.desiredHeaders, 'want more?', w
   },
 
   refresh: function() {
-    // no one should actually call this.
+    // no one should actually call this.  If they do, we absolutely don't want
+    // to do anything since we may span a sufficiently large time-range that it
+    // would be insane for our current/baseline IMAP support.  Eventually, on
+    // QRESYNC-capable IMAP and things like ActiveSync/POP3 where sync is
+    // simple it would make sense to pass this through.
+  },
+
+  /**
+   * We are hearing about a new header (possibly with body), or have transformed
+   * an onHeaderModified notification into onHeaderAdded since there's a
+   * possibility the header may now match the search filter.
+   *
+   * It is super important to keep in mind that / be aware of:
+   * - We only get called about headers that are inside the range we already
+   *   cover or if FolderStorage thinks the slice should grow because of being
+   *   latched to the top or something like that.
+   * - We maintain the start/end ranges based on the input to the filtering step
+   *   and not the filtered results.  So we always want to apply the start/end
+   *   update logic.
+   */
+  onHeaderAdded: function(header, body) {
+    if (!this._bridgeHandle || this._loading) {
+      return;
+    }
+
+    // COPY-N-PASTE: logic from MailSlice.onHeaderAdded
+    if (this.startTS === null ||
+        BEFORE(header.date, this.startTS)) {
+      this.startTS = header.date;
+      this.startUID = header.id;
+    }
+    else if (header.date === this.startTS &&
+             header.id < this.startUID) {
+      this.startUID = header.id;
+    }
+    if (this.endTS === null ||
+        STRICTLY_AFTER(header.date, this.endTS)) {
+      this.endTS = header.date;
+      this.endUID = header.id;
+    }
+    else if (header.date === this.endTS &&
+             header.id > this.endUID) {
+      this.endUID = header.id;
+    }
+    // END COPY-N-PASTE
+
+    var matchObj = this.filterer.testMessage(header, body);
+    if (!matchObj) {
+      // In the range-extending case, addMessageHeader may help us out by
+      // boosting our desiredHeaders.  It does this assuming we will then
+      // include the header like a normal slice, so we need to correct for this
+      // be capping ourselves back to desiredHeaders again
+      this.desiredHeaders = this.headers.length;
+      return;
+    }
+
+    var wrappedHeader = { header: header, matches: matchObj };
+    var idx = bsearchForInsert(this.headers, wrappedHeader,
+                               cmpMatchHeadersYoungToOld);
+
+    // We don't need to do headers.length checking here because the caller
+    // checks this for us sufficiently.  (The inclusion of the logic in
+    // MailSlice.onHeaderAdded relates to slices directly fed by the sync
+    // process which may be somewhat moot but definite is not something that
+    // happens to us, a search slice.)
+    //
+    // For sanity, we should make sure desiredHeaders doesn't get out-of-wack,
+    // though.
+    this.desiredHeaders = this.headers.length;
+
+    this._LOG.headerAdded(idx, wrappedHeader);
+    this.headers.splice(idx, 0, wrappedHeader);
+    this.headerCount = this.headers.length +
+      (this.atBottom ? 0 : this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM);
+    this._bridgeHandle.sendSplice(idx, 0, [wrappedHeader], false, false);
+  },
+
+  /**
+   * As a shortcut on many levels, we only allow messages to transition from not
+   * matching to matching.  This is logically consistent since we don't support
+   * filtering on the user-mutable aspects of a message (flags / folder / etc.),
+   * but can end up downloading more pieces of a message's body which can result
+   * in a message starting to match.
+   *
+   * This is also a correctness shortcut since we rely on body-hints to be
+   * provided by synchronization logic.  They will be provided when the body is
+   * being updated since we always update the header at the same time, but will
+   * not be provided in the case of flag-only changes.  Obviously it would suck
+   * if the flagged state of a message changed and then we dropped the message
+   * from the match list because we had no body against which to match.  There
+   * are things we could do to track body-matchingness indepenently of the flags
+   * but it's simplest to just only allow the 1-way transition for now.
+   */
+  onHeaderModified: function(header, body) {
+    if (!this._bridgeHandle || this._loading) {
+      return;
+    }
+
+
+    var wrappedHeader = { header: header, matches: null };
+    var idx = bsearchMaybeExists(this.headers, wrappedHeader,
+                                 cmpMatchHeadersYoungToOld);
+    if (idx !== null) {
+      // Update the header in the match and send it out.
+      var existingMatch = this.headers[idx];
+      existingMatch.header = header;
+      this._LOG.headerModified(idx, existingMatch);
+      this._bridgeHandle.sendUpdate([idx, existingMatch]);
+      return;
+    }
+
+    // No transition is possible if we don't care about bodies or don't have one
+    if (!this.filterer.bodiesNeeded || !body) {
+      return;
+    }
+
+    // Okay, let the add logic see if it fits.
+    this.onHeaderAdded(header, body);
+  },
+
+  onHeaderRemoved: function(header) {
+    if (!this._bridgeHandle) {
+      return;
+    }
+    // NB: We must always apply this logic since our range characterizes what we
+    // have searched/filtered, not what's inside us.  Unfortunately, when this
+    // does happen, we will drastically decrease our scope to the mesages
+    // we have matched.  What we really need for maximum correctness is to be
+    // able to know the message namers on either side of the header being
+    // deleted.  This could be interrogated by us or provided by the caller.
+    //
+    // (This would not necessitate additional block loads since if the header is
+    // at either end of its containing block, then the namer for the thing on
+    // the other side is known from the message namer defining the adjacent
+    // block.)
+    //
+    // So, TODO: Do not drastically decrease range / lose 'latch to new'
+    // semantics when the messages bounding our search get deleted.
+    //
+    // COPY-N-PASTE-N-MODIFY: logic from MailSlice.onHeaderRemoved
+    if (header.date === this.endTS && header.id === this.endUID) {
+      if (!this.headers.length) {
+        this.endTS = null;
+        this.endUID = null;
+      }
+      else {
+        this.endTS = this.headers[0].header.date;
+        this.endUID = this.headers[0].header.id;
+      }
+    }
+    if (header.date === this.startTS && header.id === this.startUID) {
+      if (!this.headers.length) {
+        this.startTS = null;
+        this.startUID = null;
+      }
+      else {
+        var lastHeader = this.headers[this.headers.length - 1];
+        this.startTS = lastHeader.header.date;
+        this.startUID = lastHeader.header.id;
+      }
+    }
+    // END COPY-N-PASTE
+
+    var wrappedHeader = { header: header, matches: null };
+    var idx = bsearchMaybeExists(this.headers, wrappedHeader,
+                                 cmpMatchHeadersYoungToOld);
+    if (idx !== null) {
+      this._LOG.headerRemoved(idx, wrappedHeader);
+      this.headers.splice(idx, 1);
+      this.headerCount = this.headers.length +
+        (this.atBottom ? 0 : this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM);
+      this._bridgeHandle.sendSplice(idx, 1, [], false, false);
+    }
   },
 
   reqNoteRanges: function(firstIndex, firstSuid, lastIndex, lastSuid) {
@@ -10429,42 +13676,84 @@ console.log('sf: willHave', willHave, 'of', this.desiredHeaders, 'want more?', w
       this.userCanGrowDownwards = false;
       var delCount = this.headers.length - lastIndex  - 1;
       this.desiredHeaders -= delCount;
-      if (!this._accumulating)
-        this._bridgeHandle.sendSplice(
-          lastIndex + 1, delCount, [],
-          // This is expected; more coming if there's a low-end splice
-          true, firstIndex > 0);
+
       this.headers.splice(lastIndex + 1, this.headers.length - lastIndex - 1);
-      var lastHeader = this.headers[lastIndex];
+      // (we are definitely not atBottom, so lie, lie, lie!)
+      this.headerCount = this.headers.length +
+        this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM;
+
+      this._bridgeHandle.sendSplice(
+        lastIndex + 1, delCount, [],
+        // This is expected; more coming if there's a low-end splice
+        true, firstIndex > 0);
+
+      var lastHeader = this.headers[lastIndex].header;
       this.startTS = lastHeader.date;
       this.startUID = lastHeader.id;
     }
     if (firstIndex > 0) {
       this.atTop = false;
       this.desiredHeaders -= firstIndex;
-      if (!this._accumulating)
-        this._bridgeHandle.sendSplice(0, firstIndex, [], true, false);
+
       this.headers.splice(0, firstIndex);
-      var firstHeader = this.headers[0];
+      this.headerCount = this.headers.length +
+        (this.atBottom ? 0 : this.IMAGINARY_MESSAGE_COUNT_WHEN_NOT_AT_BOTTOM);
+
+      this._bridgeHandle.sendSplice(0, firstIndex, [], true, false);
+
+      var firstHeader = this.headers[0].header;
       this.endTS = firstHeader.date;
       this.endUID = firstHeader.id;
     }
   },
 
-  reqGrow: function(dirMagnitude, userRequestsGrowth) {
-    if (dirMagnitude === -1) {
+  reqGrow: function(dirMagnitude, userRequestsGrowth, autoDoNotDesireMore) {
+    // If the caller is impatient and calling reqGrow on us before we are done,
+    // ignore them.  (Otherwise invariants will be violated, etc. etc.)  This
+    // is okay from an event perspective since we will definitely generate a
+    // completion notification, so the only way this could break the caller is
+    // if they maintained a counter of complete notifications to wait for.  But
+    // they cannot/must not do that since you can only ever get one of these!
+    // (And the race/confusion is inherently self-solving for naive code.)
+    if (!autoDoNotDesireMore && this._loading) {
+      return;
+    }
+
+    // Stop processing dynamic additions/modifications while this is happening.
+    this._loading = true;
+    var count;
+    if (dirMagnitude < 0) {
+      if (dirMagnitude === -1) {
+        count = $syncbase.INITIAL_FILL_SIZE;
+      }
+      else {
+        count = -dirMagnitude;
+      }
+      if (!autoDoNotDesireMore) {
+        this.desiredHeaders += count;
+      }
       this._storage.getMessagesAfterMessage(this.endTS, this.endUID,
-                                            $syncbase.INITIAL_FILL_SIZE,
+                                            count,
                                             this._gotMessages.bind(this, -1));
     }
-    else if (dirMagnitude === 1) {
+    else {
+      if (dirMagnitude <= 1) {
+        count = $syncbase.INITIAL_FILL_SIZE;
+      }
+      else {
+        count = dirMagnitude;
+      }
+      if (!autoDoNotDesireMore) {
+        this.desiredHeaders += count;
+      }
       this._storage.getMessagesBeforeMessage(this.startTS, this.startUID,
-                                             $syncbase.INITIAL_FILL_SIZE,
+                                             count,
                                              this._gotMessages.bind(this, 1));
     }
   },
 
   die: function() {
+    this._storage.dyingSlice(this);
     this._bridgeHandle = null;
     this._LOG.__die();
   },
@@ -10474,8 +13763,16 @@ var LOGFAB = exports.LOGFAB = $log.register($module, {
   SearchSlice: {
     type: $log.QUERY,
     events: {
+      headersAppended: { index: false },
+      headerAdded: { index: false },
+      headerModified: { index: false },
+      headerRemoved: { index: false },
     },
     TEST_ONLY_events: {
+      headersAppended: { headers: false },
+      headerAdded: { header: false },
+      headerModified: { header: false },
+      headerRemoved: { header: false },
     },
   },
 }); // end LOGFAB
@@ -10652,7 +13949,7 @@ exports.local_do_modtags = function(op, doneCallback, undo) {
           }
         }
         storage.updateMessageHeader(header.date, header.id, false,
-                                    header, next);
+                                    header, /* body hint */ null, next);
       }
     },
     function() {
@@ -10743,7 +14040,7 @@ exports.local_do_move = function(op, doneCallback, targetFolderId) {
 
         stateDelta.moveMap[sourceSuid] = header.suid;
         addWait = 2;
-        targetStorage.addMessageHeader(header, added);
+        targetStorage.addMessageHeader(header, body, added);
         targetStorage.addMessageBody(header, body, added);
       }
       function added() {
@@ -11975,7 +15272,8 @@ draftsMixins.local_do_saveDraft = function(op, callback) {
             op.existingNamer.suid, op.existingNamer.date, dbModCompleted);
         }
 
-        folderStorage.addMessageHeader(newRecords.header, dbModCompleted);
+        folderStorage.addMessageHeader(newRecords.header, newRecords.body,
+                                       dbModCompleted);
         folderStorage.addMessageBody(newRecords.header, newRecords.body,
                                      dbModCompleted);
         function dbModCompleted() {
@@ -13250,6 +16548,8 @@ function SliceBridgeProxy(bridge, ns, handle) {
   this.progress = 0.0;
   this.atTop = false;
   this.atBottom = false;
+  this.headerCount = 0;
+
   /**
    * Can we potentially grow the slice in the negative direction if explicitly
    * desired by the user or UI desires to be up-to-date?  For example,
@@ -13288,7 +16588,14 @@ SliceBridgeProxy.prototype = {
       requested: requested,
       moreExpected: moreExpected,
       newEmailCount: newEmailCount,
-      type: 'slice',
+      // send header count here instead of batchSlice,
+      // since need an accurate count for each splice
+      // call: there could be two splices for the 0
+      // index in a row, and setting count on the
+      // batchSlice will not give an accurate picture
+      // that the slice actions are growing the slice.
+      headerCount: this.headerCount,
+      type: 'slice'
     };
     this.addUpdate(updateSplice);
   },
